@@ -6,7 +6,7 @@ version: 1.0.43+49
 #flutter version 3.7.9;
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
