@@ -3,13 +3,13 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 abstract class AssortmentFiltersEvent {}
 
 class AssortmentFiltersLoadEvent extends AssortmentFiltersEvent {
-  String catalogUuid;
-  bool isFavorite;
-  bool isYellowTags;
-  int currentPage;
-  List<String> activeTagsList;
-  String searchText;
-  bool isBranded;
+  String? catalogUuid;
+  bool? isFavorite;
+  bool? isYellowTags;
+  int? currentPage;
+  List<String>? activeTagsList;
+  String? searchText;
+  bool? isBranded;
 
   AssortmentFiltersLoadEvent({
     this.catalogUuid,
@@ -25,13 +25,13 @@ class AssortmentFiltersLoadEvent extends AssortmentFiltersEvent {
 abstract class AssortmentFiltersEState {}
 
 class AssortmentFiltersActiveState extends AssortmentFiltersEState {
-  String catalogUuid;
-  bool isFavorite;
-  bool isYellowTags;
-  int currentPage;
-  List<String> activeTagsList;
-  String searchText;
-  bool isBranded;
+  String? catalogUuid;
+  bool? isFavorite;
+  bool? isYellowTags;
+  int? currentPage;
+  List<String>? activeTagsList;
+  String? searchText;
+  bool? isBranded;
 
   AssortmentFiltersActiveState({
     this.catalogUuid,
@@ -46,11 +46,13 @@ class AssortmentFiltersActiveState extends AssortmentFiltersEState {
 
 class AssortmentFiltersDisableState extends AssortmentFiltersEState {}
 
-class AssortmentFiltersBloc extends Bloc<AssortmentFiltersEvent, AssortmentFiltersEState> {
+class AssortmentFiltersBloc
+    extends Bloc<AssortmentFiltersEvent, AssortmentFiltersEState> {
   AssortmentFiltersBloc() : super(AssortmentFiltersDisableState());
 
   @override
-  Stream<AssortmentFiltersEState> mapEventToState(AssortmentFiltersEvent event) async* {
+  Stream<AssortmentFiltersEState> mapEventToState(
+      AssortmentFiltersEvent event) async* {
     if (event is AssortmentFiltersLoadEvent) {
       yield AssortmentFiltersActiveState(
         activeTagsList: event.activeTagsList,
