@@ -1,6 +1,4 @@
 //events
-
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/credit_cards_list_model.dart';
 import 'package:smart/services/services.dart';
@@ -22,19 +20,22 @@ class CreditCardsListErrorState extends CreditCardsListState {}
 class CreditCardsListLoadedState extends CreditCardsListState {
   final CreditCardsListModel cardsListModel;
 
-  CreditCardsListLoadedState({@required this.cardsListModel});
+  CreditCardsListLoadedState({required this.cardsListModel});
 }
 
 //bloc class
 
-class CreditCardsListBloc extends Bloc<CreditCardsListEvent, CreditCardsListState> {
+class CreditCardsListBloc
+    extends Bloc<CreditCardsListEvent, CreditCardsListState> {
   CreditCardsListBloc() : super(CreditCardsListInitState());
 
   @override
-  Stream<CreditCardsListState> mapEventToState(CreditCardsListEvent event) async* {
+  Stream<CreditCardsListState> mapEventToState(
+      CreditCardsListEvent event) async* {
     yield CreditCardsListLoadingState();
     try {
-      final _creditCardsListModel = await CreditCardsProvider().getCreditCardsListResponce();
+      final _creditCardsListModel =
+          await CreditCardsProvider().getCreditCardsListResponce();
       yield CreditCardsListLoadedState(cardsListModel: _creditCardsListModel);
     } catch (e) {
       yield CreditCardsListErrorState();
