@@ -1,5 +1,4 @@
 //events
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/onboarding_list_Model.dart';
@@ -10,7 +9,7 @@ abstract class OnboardingEvent {}
 class OnboardingLoadEvent extends OnboardingEvent {
   final BuildContext context;
 
-  OnboardingLoadEvent({@required this.context});
+  OnboardingLoadEvent({required this.context});
 }
 
 //state
@@ -25,7 +24,7 @@ class OnboardingErrorstate extends OnboardingState {}
 class OnboardingLoadedstate extends OnboardingState {
   final OnboardingListModel onboardingListModel;
 
-  OnboardingLoadedstate({@required this.onboardingListModel});
+  OnboardingLoadedstate({required this.onboardingListModel});
 }
 
 //bloc class
@@ -37,7 +36,8 @@ class OnboardingBloc extends Bloc<OnboardingEvent, OnboardingState> {
     if (event is OnboardingLoadEvent) {
       yield OnboardingLoadingstate();
       try {
-        OnboardingListModel _onboardingListModel = await OnboardingListProvider().getOnboradingListResponse();
+        OnboardingListModel _onboardingListModel =
+            await OnboardingListProvider().getOnboradingListResponse();
         // for (var i = 0; i < _onboardingListModel.data.length; i++) {
         //   await UtilsForOnboard.cachImage(
         //       event.context, _onboardingListModel.data[i].logoFilePath);
@@ -46,7 +46,8 @@ class OnboardingBloc extends Bloc<OnboardingEvent, OnboardingState> {
         if (_onboardingListModel.data.isEmpty) {
           yield OnboardingErrorstate();
         } else {
-          yield OnboardingLoadedstate(onboardingListModel: _onboardingListModel);
+          yield OnboardingLoadedstate(
+              onboardingListModel: _onboardingListModel);
         }
       } catch (error) {
         yield OnboardingErrorstate();
