@@ -1,5 +1,4 @@
 import 'dart:developer';
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/order_calculate_response_model.dart';
 import 'package:smart/models/product_model_for_order_request.dart';
@@ -8,17 +7,21 @@ import 'package:smart/services/services.dart';
 abstract class OrderCalculateEvent {}
 
 class OrderCalculateLoadEvent extends OrderCalculateEvent {
-  final String orderDeliveryTypeId;
-  final String orderPaymentTypeId;
-  final List<ProductModelForOrderRequest> productModelForOrderRequestList;
-  final int subtractBonusesCount;
-  OrderCalculateLoadEvent(
-      {this.subtractBonusesCount, @required this.orderDeliveryTypeId, @required this.orderPaymentTypeId, @required this.productModelForOrderRequestList});
+  final String? orderDeliveryTypeId;
+  final String? orderPaymentTypeId;
+  final List<ProductModelForOrderRequest>? productModelForOrderRequestList;
+  final int? subtractBonusesCount;
+  OrderCalculateLoadEvent({
+    this.subtractBonusesCount,
+    required this.orderDeliveryTypeId,
+    required this.orderPaymentTypeId,
+    required this.productModelForOrderRequestList,
+  });
 }
 
 //states
 abstract class OrderCalculateState {
-  final OrderDetailsAndCalculateResponseModel orderCalculateResponseModel;
+  final OrderDetailsAndCalculateResponseModel? orderCalculateResponseModel;
   OrderCalculateState(this.orderCalculateResponseModel);
 }
 
@@ -27,7 +30,9 @@ class OrderCalculateInitState extends OrderCalculateState {
 }
 
 class OrderCalculateLoadingState extends OrderCalculateState {
-  OrderCalculateLoadingState({OrderDetailsAndCalculateResponseModel orderCalculateResponseModel}) : super(orderCalculateResponseModel);
+  OrderCalculateLoadingState({
+    OrderDetailsAndCalculateResponseModel? orderCalculateResponseModel,
+  }) : super(orderCalculateResponseModel);
 }
 
 class OrderCalculateErrorState extends OrderCalculateState {
@@ -35,14 +40,17 @@ class OrderCalculateErrorState extends OrderCalculateState {
 }
 
 class OrderCalculateLoadedState extends OrderCalculateState {
-  OrderCalculateLoadedState({OrderDetailsAndCalculateResponseModel orderCalculateResponseModel}) : super(orderCalculateResponseModel);
+  OrderCalculateLoadedState({
+    OrderDetailsAndCalculateResponseModel? orderCalculateResponseModel,
+  }) : super(orderCalculateResponseModel);
 }
 
-class OrderCalculateBloc extends Bloc<OrderCalculateEvent, OrderCalculateState> {
+class OrderCalculateBloc
+    extends Bloc<OrderCalculateEvent, OrderCalculateState> {
   OrderCalculateBloc() : super(OrderCalculateInitState());
-  OrderDetailsAndCalculateResponseModel _orderDetailsAndCalculateResponseModel;
-  String _promocode;
-  String get promocode => _promocode;
+  OrderDetailsAndCalculateResponseModel? _orderDetailsAndCalculateResponseModel;
+  String? _promocode;
+  String? get promocode => _promocode;
 
   void updatePromoCode(String newPromoCode) {
     _promocode = newPromoCode;
@@ -50,28 +58,35 @@ class OrderCalculateBloc extends Bloc<OrderCalculateEvent, OrderCalculateState> 
   }
 
   @override
-  Stream<OrderCalculateState> mapEventToState(OrderCalculateEvent event) async* {
+  Stream<OrderCalculateState> mapEventToState(
+      OrderCalculateEvent event) async* {
     if (event is OrderCalculateLoadEvent) {
-      yield OrderCalculateLoadingState(orderCalculateResponseModel: _orderDetailsAndCalculateResponseModel);
+      yield OrderCalculateLoadingState(
+          orderCalculateResponseModel: _orderDetailsAndCalculateResponseModel);
       try {
-        if (event.productModelForOrderRequestList.isNotEmpty) {
+        if (event.productModelForOrderRequestList?.isNotEmpty == true) {
           log('🔄 Отправка запроса на расчёт заказа с промокодом: $_promocode');
-          _orderDetailsAndCalculateResponseModel = await OrderProvider().orderCalculateResponse(
+          _orderDetailsAndCalculateResponseModel =
+              await OrderProvider().orderCalculateResponse(
             subtractBonusesCount: event.subtractBonusesCount,
             orderDeliveryTypeId: event.orderDeliveryTypeId,
             orderPaymentTypeId: event.orderPaymentTypeId,
-            productModelForOrderRequestList: event.productModelForOrderRequestList,
+            productModelForOrderRequestList:
+                event.productModelForOrderRequestList,
             promocode: _promocode,
           );
           //
           if (_orderDetailsAndCalculateResponseModel?.data?.promocode != null) {
-            _promocode = _orderDetailsAndCalculateResponseModel.data.promocode;
+            _promocode =
+                _orderDetailsAndCalculateResponseModel?.data?.promocode;
             log('✅ Промокод успешно применён: $_promocode');
           } else {
             _promocode = null;
             log('❌ Промокод невалидный или устарел');
           }
-          yield OrderCalculateLoadedState(orderCalculateResponseModel: _orderDetailsAndCalculateResponseModel);
+          yield OrderCalculateLoadedState(
+              orderCalculateResponseModel:
+                  _orderDetailsAndCalculateResponseModel);
         } else {
           log('⚠️ Отправка загрузки счета ⚠️ на калькуляцию запрос не отправлен так как список с продуктами пуст');
         }
