@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/product_details_model.dart';
 
@@ -9,24 +8,24 @@ abstract class ProductDetMiddleContentEvent {}
 class ProductDetMiddleContentDescriptionEvent
     extends ProductDetMiddleContentEvent {
   final ProductDetailsModel productDetailsModel;
-  ProductDetMiddleContentDescriptionEvent({@required this.productDetailsModel});
+  ProductDetMiddleContentDescriptionEvent({required this.productDetailsModel});
 }
 
 class ProductDetMiddleContentIngredientsEvent
     extends ProductDetMiddleContentEvent {
   final ProductDetailsModel productDetailsModel;
 
-  ProductDetMiddleContentIngredientsEvent({@required this.productDetailsModel});
+  ProductDetMiddleContentIngredientsEvent({required this.productDetailsModel});
 }
 
 class ProductDetMiddleContentComentsEvent extends ProductDetMiddleContentEvent {
   final ProductDetailsModel productDetailsModel;
 
-  ProductDetMiddleContentComentsEvent({@required this.productDetailsModel});
+  ProductDetMiddleContentComentsEvent({required this.productDetailsModel});
 }
 
-class ProductDetMiddleContentDisableEvent extends ProductDetMiddleContentEvent {
-}
+class ProductDetMiddleContentDisableEvent
+    extends ProductDetMiddleContentEvent {}
 
 //states
 
@@ -36,27 +35,27 @@ class ProductDetMiddleContentEmptyState extends ProductDetMiddleContentState {}
 
 class ProductDetMiddleContentErrorState extends ProductDetMiddleContentState {}
 
-class ProductDetMiddleContentLoadingState extends ProductDetMiddleContentState {
-}
+class ProductDetMiddleContentLoadingState
+    extends ProductDetMiddleContentState {}
 
 class ProductDetMiddleContentDescriptionState
     extends ProductDetMiddleContentState {
   final ProductDetailsModel productDetailsModel;
 
-  ProductDetMiddleContentDescriptionState({@required this.productDetailsModel});
+  ProductDetMiddleContentDescriptionState({required this.productDetailsModel});
 }
 
 class ProductDetMiddleContentIngredientsState
     extends ProductDetMiddleContentState {
   final ProductDetailsModel productDetailsModel;
 
-  ProductDetMiddleContentIngredientsState({@required this.productDetailsModel});
+  ProductDetMiddleContentIngredientsState({required this.productDetailsModel});
 }
 
 class ProductDetMiddleContentComentsState extends ProductDetMiddleContentState {
   final ProductDetailsModel productDetailsModel;
 
-  ProductDetMiddleContentComentsState({@required this.productDetailsModel});
+  ProductDetMiddleContentComentsState({required this.productDetailsModel});
 }
 
 class ProductDetMiddleContentBloc
