@@ -1,7 +1,5 @@
 //events
 import 'dart:async';
-
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/product_details_data_model.dart';
 
@@ -11,16 +9,20 @@ class ProductInShopAsMapEvent extends ProductInShopEvent {
   final List<ProductDetailsStoreListModel> storesListModel;
   final String assortmentUnitId;
 
-  ProductInShopAsMapEvent(
-      {@required this.assortmentUnitId, @required this.storesListModel});
+  ProductInShopAsMapEvent({
+    required this.assortmentUnitId,
+    required this.storesListModel,
+  });
 }
 
 class ProductInShopAsListEvent extends ProductInShopEvent {
   final List<ProductDetailsStoreListModel> storesListModel;
   final String assortmentUnitId;
 
-  ProductInShopAsListEvent(
-      {@required this.assortmentUnitId, @required this.storesListModel});
+  ProductInShopAsListEvent({
+    required this.assortmentUnitId,
+    required this.storesListModel,
+  });
 }
 
 //states
@@ -33,15 +35,19 @@ class ProductInShopAsMapState extends ProductInShopState {
   final List<ProductDetailsStoreListModel> storesListModel;
   final String assortmentUnitId;
 
-  ProductInShopAsMapState(
-      {@required this.assortmentUnitId, @required this.storesListModel});
+  ProductInShopAsMapState({
+    required this.assortmentUnitId,
+    required this.storesListModel,
+  });
 }
 
 class ProductInShopAsListState extends ProductInShopState {
   final List<ProductDetailsStoreListModel> storesListModel;
   final String assortmentUnitId;
-  ProductInShopAsListState(
-      {@required this.assortmentUnitId, @required this.storesListModel});
+  ProductInShopAsListState({
+    required this.assortmentUnitId,
+    required this.storesListModel,
+  });
 }
 
 class ProductInShopBloc extends Bloc<ProductInShopEvent, ProductInShopState> {
