@@ -2,7 +2,6 @@
 
 import 'dart:developer';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -31,7 +30,11 @@ class SmartContactsLoadedState extends SmartContactsState {
   final SocialsListModel socialsListModel;
   final String appVersion;
 
-  SmartContactsLoadedState({@required this.socialsListModel, @required this.smartContactsModel, @required this.appVersion});
+  SmartContactsLoadedState({
+    required this.socialsListModel,
+    required this.smartContactsModel,
+    required this.appVersion,
+  });
 }
 
 //bloc class
@@ -43,18 +46,26 @@ class SmartContactsBloc extends Bloc<SmartContactsEvent, SmartContactsState> {
   Stream<SmartContactsState> mapEventToState(SmartContactsEvent event) async* {
     yield SmartContactsLoadinState();
     try {
-      SmartContactsModel _smartContactsModel = await SmartContactsProvider().getContactsResponse();
+      SmartContactsModel _smartContactsModel =
+          await SmartContactsProvider().getContactsResponse();
       prefs = await SharedPreferences.getInstance();
-      await prefs.setString(SharedKeys.callCenterNumber, _smartContactsModel.data.callCenterNumber);
+      await prefs.setString(SharedKeys.callCenterNumber,
+          _smartContactsModel.data.callCenterNumber);
       log('попап отображать: ');
-      await prefs.setBool(SharedKeys.isSocialNetworkFacebook, _smartContactsModel.data.socialNetworkFacebook == '1' ? true : false);
+      await prefs.setBool(SharedKeys.isSocialNetworkFacebook,
+          _smartContactsModel.data.socialNetworkFacebook == '1' ? true : false);
       log('попап отображать: ${prefs.getBool(SharedKeys.isSocialNetworkFacebook)}');
 
-      SocialsListModel _socialsListModel = await SmartContactsProvider().getSocialsListResponse();
-      await prefs.setString(SharedKeys.socialsListModelUrlWhatsapp, _socialsListModel.data[0].url);
+      SocialsListModel _socialsListModel =
+          await SmartContactsProvider().getSocialsListResponse();
+      await prefs.setString(SharedKeys.socialsListModelUrlWhatsapp,
+          _socialsListModel.data[0].url);
 
       PackageInfo packageInfo = await PackageInfo.fromPlatform();
-      yield SmartContactsLoadedState(appVersion: packageInfo.version, smartContactsModel: _smartContactsModel, socialsListModel: _socialsListModel);
+      yield SmartContactsLoadedState(
+          appVersion: packageInfo.version,
+          smartContactsModel: _smartContactsModel,
+          socialsListModel: _socialsListModel);
     } catch (_) {
       yield SmartContactsErrorState();
     }
