@@ -1,5 +1,4 @@
 //event
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/stories_list_model.dart';
 import 'package:smart/services/services.dart';
@@ -19,7 +18,7 @@ class StoriesListInitState extends StoriesListState {}
 
 class StoriesListLoadedState extends StoriesListState {
   final StoriesListModel storiesListModel;
-  StoriesListLoadedState({@required this.storiesListModel});
+  StoriesListLoadedState({required this.storiesListModel});
 }
 
 //BLoC class
@@ -30,7 +29,8 @@ class StoriesListBloc extends Bloc<StoriesListEvent, StoriesListState> {
   Stream<StoriesListState> mapEventToState(StoriesListEvent event) async* {
     yield StoriesListLoadingState();
     try {
-      StoriesListModel _storiesListModel = await StoriesProvider().storiesResponse();
+      StoriesListModel _storiesListModel =
+          await StoriesProvider().storiesResponse();
 
       if (_storiesListModel.data.isEmpty) {
         yield StoriesListInitState();
