@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart/models/assortments_list_model.dart';
 import 'package:smart/services/services.dart';
@@ -21,7 +20,7 @@ class YellowPromoAssortmentsEmptyState extends YellowPromoAssortmentsState {}
 class YellowPromoAssortmentsLoadedState extends YellowPromoAssortmentsState {
   final List<AssortmentsListModel> assortmentsList;
 
-  YellowPromoAssortmentsLoadedState({@required this.assortmentsList});
+  YellowPromoAssortmentsLoadedState({required this.assortmentsList});
 }
 
 class YellowPromoAssortmentsBloc
