@@ -86,10 +86,12 @@ class OrderProcessProvider {
       Uri.parse(_orderProcessListUrl),
       headers: {"Accept": "application/json"},
     );
-    log('url:: get Future<OrderProcessListModel> orderProcessResponse ' + _orderProcessListUrl);
+    log('url:: get Future<OrderProcessListModel> orderProcessResponse ' +
+        _orderProcessListUrl);
     print('✅ ${respone.body}');
     if (respone.statusCode == 200) {
-      final Map<String, dynamic> orderProcessListJson = json.decode(respone.body);
+      final Map<String, dynamic> orderProcessListJson =
+          json.decode(respone.body);
       return OrderProcessListModel.fromJson(orderProcessListJson);
     } else {
       throw Exception("Ошибка. ❌ Данные о заказах не прогружены");
@@ -123,11 +125,13 @@ class VacancyListProvider {
 class BonusesListProvider {
   BonusesListProvider();
 
-  Future<List<BonusesListDataModel>> BonusesListForPaginationResponse({@required int currentPage}) async {
+  Future<List<BonusesListDataModel>> BonusesListForPaginationResponse(
+      {@required int currentPage}) async {
     String token = await loadToken();
     String url =
         "$apiHead/clients/api/profile/client-bonus-transactions?api_token=$token&order_by[created_at]=desc&page=$currentPage";
-    final response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    final response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
     log('get ' + url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> bonusesListJson = json.decode(response.body);
@@ -151,7 +155,8 @@ class OnboardingListProvider {
     );
     log('getOnboradingListResponse ' + _onboardingListUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _onboardingListJson = json.decode(response.body);
+      final Map<String, dynamic> _onboardingListJson =
+          json.decode(response.body);
       return OnboardingListModel.fromJson(_onboardingListJson);
     } else {
       throw Exception("error  fetching onborading list!");
@@ -184,7 +189,8 @@ class StoriesProvider {
 
   Future<StoriesListModel> storiesResponse() async {
     String token = await loadToken();
-    _storiesListUrl = "$apiHead/clients/api/stories?order_by[created_at]=asc&api_token=$token&per_page=1000";
+    _storiesListUrl =
+        "$apiHead/clients/api/stories?order_by[created_at]=asc&api_token=$token&per_page=1000";
     final respone = await http.get(
       Uri.parse(_storiesListUrl),
       headers: {"Accept": "application/json"},
@@ -217,11 +223,12 @@ class AddressesClientProvider {
     );
     if (response.statusCode == 200) {
       final Map<String, dynamic> clientAddressJson = json.decode(response.body);
-      AddressesClientListModel model = AddressesClientListModel.fromJson(clientAddressJson);
+      AddressesClientListModel model =
+          AddressesClientListModel.fromJson(clientAddressJson);
       print('ClientAddressModel:::: ${model.data}');
       if (model.data.isNotEmpty) {
-        await prefs.setString(
-            SharedKeys.myAddress, '${model.data[0].city}, ${model.data[0].street} ${model.data[0].house}');
+        await prefs.setString(SharedKeys.myAddress,
+            '${model.data[0].city}, ${model.data[0].street} ${model.data[0].house}');
         await prefs.setString(SharedKeys.myAddressUuid, model.data[0].uuid);
       }
       return model;
@@ -250,18 +257,24 @@ class AddressesClientProvider {
       if (street != null && street.isNotEmpty) "street": street,
       if (house != null && house.isNotEmpty) "house": house,
       if (floor != null && floor.isNotEmpty) "floor": floor,
-      if (apartmentNumber != null && apartmentNumber.isNotEmpty) "apartment_number": apartmentNumber,
-      if (intercomCode != null && intercomCode.isNotEmpty) "intercom_code": intercomCode,
+      if (apartmentNumber != null && apartmentNumber.isNotEmpty)
+        "apartment_number": apartmentNumber,
+      if (intercomCode != null && intercomCode.isNotEmpty)
+        "intercom_code": intercomCode,
     });
-    log('Cохраняем адрес клиента 🏠 post addClientAddressResponse: ' + _clienAddressUrl);
+    log('Cохраняем адрес клиента 🏠 post addClientAddressResponse: ' +
+        _clienAddressUrl);
     if (title != null && title.isNotEmpty) print('title👉 ' + title);
-    if (entrance != null && entrance.isNotEmpty) print('entrance:: ' + entrance);
+    if (entrance != null && entrance.isNotEmpty)
+      print('entrance:: ' + entrance);
     if (city != null && city.isNotEmpty) print('city:: ' + city);
     if (street != null && street.isNotEmpty) print('street:: ' + street);
     if (house != null && house.isNotEmpty) print('house:: ' + house);
     if (floor != null && floor.isNotEmpty) print('floor:: ' + floor);
-    if (apartmentNumber != null && apartmentNumber.isNotEmpty) print('apartment_number:: ' + apartmentNumber);
-    if (intercomCode != null && intercomCode.isNotEmpty) print('intercom_code:: ' + intercomCode);
+    if (apartmentNumber != null && apartmentNumber.isNotEmpty)
+      print('apartment_number:: ' + apartmentNumber);
+    if (intercomCode != null && intercomCode.isNotEmpty)
+      print('intercom_code:: ' + intercomCode);
     if (response.statusCode == 201) {
       // final Map<String, dynamic> clientAddressJson = json.decode(response.body);
       return true;
@@ -292,25 +305,42 @@ class AddressesClientProvider {
     }, body: {
       if (title != null && title != "null" && title.isNotEmpty) "title": title,
       if (city != null && city != "null" && city.isNotEmpty) "city": city,
-      if (street != null && street != "null" && street.isNotEmpty) "street": street,
+      if (street != null && street != "null" && street.isNotEmpty)
+        "street": street,
       if (house != null && house != "null" && house.isNotEmpty) "house": house,
       if (floor != null && floor != "null" && floor.isNotEmpty) "floor": floor,
-      if (apartmentNumber != null && apartmentNumber != "null" && apartmentNumber.isNotEmpty)
+      if (apartmentNumber != null &&
+          apartmentNumber != "null" &&
+          apartmentNumber.isNotEmpty)
         "apartment_number": apartmentNumber,
-      if (entrance != null && entrance != "null" && entrance.isNotEmpty) "entrance": entrance,
-      if (intercomCode != null && intercomCode != "null" && intercomCode.isNotEmpty) "intercom_code": intercomCode,
+      if (entrance != null && entrance != "null" && entrance.isNotEmpty)
+        "entrance": entrance,
+      if (intercomCode != null &&
+          intercomCode != "null" &&
+          intercomCode.isNotEmpty)
+        "intercom_code": intercomCode,
     });
-    log('Обновляем адрес клиента 🏠♻️ put changeClientAddressResponse: ' + _clienAddressUrl);
-    if (title != null && title != "null" && title.isNotEmpty) print('title👉 ' + title);
-    if (city != null && city != "null" && city.isNotEmpty) print('city:: ' + city);
-    if (street != null && street != "null" && street.isNotEmpty) print('street:: ' + street);
-    if (house != null && house != "null" && house.isNotEmpty) print('house:: ' + house);
-    if (floor != null && floor != "null" && floor.isNotEmpty) print('floor:: ' + floor);
-    if (apartmentNumber != null && apartmentNumber != "null" && apartmentNumber.isNotEmpty)
+    log('Обновляем адрес клиента 🏠♻️ put changeClientAddressResponse: ' +
+        _clienAddressUrl);
+    if (title != null && title != "null" && title.isNotEmpty)
+      print('title👉 ' + title);
+    if (city != null && city != "null" && city.isNotEmpty)
+      print('city:: ' + city);
+    if (street != null && street != "null" && street.isNotEmpty)
+      print('street:: ' + street);
+    if (house != null && house != "null" && house.isNotEmpty)
+      print('house:: ' + house);
+    if (floor != null && floor != "null" && floor.isNotEmpty)
+      print('floor:: ' + floor);
+    if (apartmentNumber != null &&
+        apartmentNumber != "null" &&
+        apartmentNumber.isNotEmpty)
       print('apartment_number:: ' + apartmentNumber);
-    if (entrance != null && entrance != "null" && entrance.isNotEmpty) print('entrance:: ' + entrance);
-    if (intercomCode != null && intercomCode != "null" && intercomCode.isNotEmpty)
-      print('intercom_code:: ' + intercomCode);
+    if (entrance != null && entrance != "null" && entrance.isNotEmpty)
+      print('entrance:: ' + entrance);
+    if (intercomCode != null &&
+        intercomCode != "null" &&
+        intercomCode.isNotEmpty) print('intercom_code:: ' + intercomCode);
     if (response.statusCode == 200) {
       // final Map<String, dynamic> clientAddressJson = json.decode(response.body);
       return true;
@@ -323,8 +353,10 @@ class AddressesClientProvider {
   Future<bool> deleteClientAddressResponse({@required String addressId}) async {
     String token = await loadToken();
     _clienAddressUrl += "/{$addressId}?api_token=$token";
-    final response = await http.delete(Uri.parse(_clienAddressUrl), headers: {"Accept": "application/json"});
-    log('удаляем адрес клиента 🏠 delete ClientAddressResponse ' + _clienAddressUrl);
+    final response = await http.delete(Uri.parse(_clienAddressUrl),
+        headers: {"Accept": "application/json"});
+    log('удаляем адрес клиента 🏠 delete ClientAddressResponse ' +
+        _clienAddressUrl);
     if (response.statusCode == 200) {
       // final Map<String, dynamic> clientAddressJson = json.decode(response.body);
       return true;
@@ -339,16 +371,20 @@ class AssortmentBrandsProvider {
   //TODO1 heart
   String _assortBrandsUrl = "$apiHead/clients/api/assortment-brands";
 
-  Future<AssortmentBrandsListmodel> getAssortmentBrandsResponse({@required int page, String searchText}) async {
+  Future<AssortmentBrandsListmodel> getAssortmentBrandsResponse(
+      {@required int page, String searchText}) async {
     String token = await loadToken();
     _assortBrandsUrl += "?api_token=$token&page=$page&order_by[name] = asc";
     if (searchText != null) {
-      _assortBrandsUrl += "&where[0][0]=name&where[0][1]=ilike&where[0][2]=%$searchText%";
+      _assortBrandsUrl +=
+          "&where[0][0]=name&where[0][1]=ilike&where[0][2]=%$searchText%";
     }
-    final response = await http.get(Uri.parse(_assortBrandsUrl), headers: {"Accept": "application/json"});
+    final response = await http.get(Uri.parse(_assortBrandsUrl),
+        headers: {"Accept": "application/json"});
     log('getAssortmentBrandsResponse get ' + _assortBrandsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _assortmentBrandsListJson = json.decode(response.body);
+      final Map<String, dynamic> _assortmentBrandsListJson =
+          json.decode(response.body);
       return AssortmentBrandsListmodel.fromJson(_assortmentBrandsListJson);
     } else {
       throw Exception("error to fetching brands list");
@@ -372,7 +408,8 @@ class ImInShopProvider {
       final Map<String, dynamic> _iminShopJson = json.decode(response.body);
       return ImInShopModel.fromJson(_iminShopJson);
     } else {
-      throw Exception("error to fetching im in shop list. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error to fetching im in shop list. StatusCode: ${response.statusCode}");
     }
   }
 
@@ -383,7 +420,8 @@ class ImInShopProvider {
       return "notStoreChoose";
     }
     final response = await http.post(Uri.parse(imInShopUrl),
-        body: {"store_uuid": shopUuid, "api_token": token}, headers: {"Accept": "application/json"});
+        body: {"store_uuid": shopUuid, "api_token": token},
+        headers: {"Accept": "application/json"});
     log('post turnOnImInShop:::::::::: ' + imInShopUrl);
     print('store_uuid:: ' + shopUuid);
     print('api_token:: ' + token);
@@ -403,10 +441,12 @@ class CreditCardsProvider {
   Future<CreditCardsListModel> getCreditCardsListResponce() async {
     String token = await loadToken();
     creditCardsUrl += "?api_token=$token";
-    final response = await http.get(Uri.parse(creditCardsUrl), headers: {"Accept": "application/json"});
+    final response = await http.get(Uri.parse(creditCardsUrl),
+        headers: {"Accept": "application/json"});
     log('💳 get:: ' + creditCardsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _creditCardsListJson = json.decode(response.body);
+      final Map<String, dynamic> _creditCardsListJson =
+          json.decode(response.body);
       return CreditCardsListModel.fromJson(_creditCardsListJson);
     } else if (response.statusCode == 401) {
       return CreditCardsListModel();
@@ -418,7 +458,8 @@ class CreditCardsProvider {
   Future<bool> deleteCreditCardResponse({@required String cardUuid}) async {
     String token = await loadToken();
     creditCardsUrl += "/{$cardUuid}?api_token=$token";
-    final response = await http.delete(Uri.parse(creditCardsUrl), headers: {"Accept": "application/json"});
+    final response = await http.delete(Uri.parse(creditCardsUrl),
+        headers: {"Accept": "application/json"});
     log('💳 delete:: ' + creditCardsUrl);
     if (response.statusCode == 200) {
       return true;
@@ -430,20 +471,24 @@ class CreditCardsProvider {
   Future<UrlForCreditCardLinkModel> getUrlForCreditCardLinkResponse() async {
     String token = await loadToken();
     creditCardsUrl += "/link?api_token=$token";
-    final response = await http.get(Uri.parse(creditCardsUrl), headers: {"Accept": "application/json"});
+    final response = await http.get(Uri.parse(creditCardsUrl),
+        headers: {"Accept": "application/json"});
     log('💳 get:: ' + creditCardsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _urlForCreditCardLinkJson = json.decode(response.body);
+      final Map<String, dynamic> _urlForCreditCardLinkJson =
+          json.decode(response.body);
       return UrlForCreditCardLinkModel.fromJson(_urlForCreditCardLinkJson);
     } else {
       throw Exception("error fetching url for card ");
     }
   }
 
-  Future<bool> setSuccessStatusOfLinkingCardResponse({@required String orderId}) async {
+  Future<bool> setSuccessStatusOfLinkingCardResponse(
+      {@required String orderId}) async {
     String token = await loadToken();
     creditCardsUrl += "/link/success?api_token=$token&orderId=$orderId";
-    final response = await http.get(Uri.parse(creditCardsUrl), headers: {"Accept": "application/json"});
+    final response = await http.get(Uri.parse(creditCardsUrl),
+        headers: {"Accept": "application/json"});
     log('💳 get setSuccessStatusOfLinkingCardResponse:: ' + creditCardsUrl);
     if (response.statusCode == 204) {
       return true;
@@ -452,10 +497,12 @@ class CreditCardsProvider {
     }
   }
 
-  Future<bool> setSErrorStatusOfLinkingCardResponse({@required String orderId}) async {
+  Future<bool> setSErrorStatusOfLinkingCardResponse(
+      {@required String orderId}) async {
     String token = await loadToken();
     creditCardsUrl += "/link/error?api_token=$token&orderId=$orderId";
-    final response = await http.get(Uri.parse(creditCardsUrl), headers: {"Accept": "application/json"});
+    final response = await http.get(Uri.parse(creditCardsUrl),
+        headers: {"Accept": "application/json"});
     log('💳 get setSErrorStatusOfLinkingCardResponse:: ' + creditCardsUrl);
     if (response.statusCode == 204) {
       return true;
@@ -468,26 +515,34 @@ class CreditCardsProvider {
 //Описание акций
 class PromoDescriptionsProvider {
   String promoDescriptionsUrl = "$apiHead/clients/api/promo-descriptions";
-  Future<List<PromoDescriptionsDataModel>> getPromoDescriptionsResponse(int page) async {
+  Future<List<PromoDescriptionsDataModel>> getPromoDescriptionsResponse(
+      int page) async {
     String token = await loadToken();
-    promoDescriptionsUrl += "?where[0][0]=is_hidden&where[0][1]==&where[0][2]=false&api_token=$token&page=$page";
-    final response = await http.get(Uri.parse(promoDescriptionsUrl), headers: {"Accept": "application/json"});
+    promoDescriptionsUrl +=
+        "?where[0][0]=is_hidden&where[0][1]==&where[0][2]=false&api_token=$token&page=$page";
+    final response = await http.get(Uri.parse(promoDescriptionsUrl),
+        headers: {"Accept": "application/json"});
     log('get getPromoDescriptionsResponse ::::: ' + promoDescriptionsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _promoDescriptionsJson = json.decode(response.body);
+      final Map<String, dynamic> _promoDescriptionsJson =
+          json.decode(response.body);
       return PromoDescriptionsModel.fromJson(_promoDescriptionsJson).data;
     } else {
       throw Exception("error fetching promo descriptions");
     }
   }
 
-  Future<PromoDescriptionsDataModel> getDiverseFoodDescriptionsResponse() async {
+  Future<PromoDescriptionsDataModel>
+      getDiverseFoodDescriptionsResponse() async {
     String token = await loadToken();
-    promoDescriptionsUrl += "?where[0][0]=name&where[0][1]==&where[0][2]=Разнообразное питание&api_token=$token";
-    final response = await http.get(Uri.parse(promoDescriptionsUrl), headers: {"Accept": "application/json"});
+    promoDescriptionsUrl +=
+        "?where[0][0]=name&where[0][1]==&where[0][2]=Разнообразное питание&api_token=$token";
+    final response = await http.get(Uri.parse(promoDescriptionsUrl),
+        headers: {"Accept": "application/json"});
     log('get getDiverseFoodDescriptionsResponse :::: ' + promoDescriptionsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _promoDescriptionsJson = json.decode(response.body);
+      final Map<String, dynamic> _promoDescriptionsJson =
+          json.decode(response.body);
       return PromoDescriptionsModel.fromJson(_promoDescriptionsJson).data.first;
     } else {
       throw Exception("error fetching promo descriptions");
@@ -496,14 +551,18 @@ class PromoDescriptionsProvider {
 
   Future<FavoriteProductTitleModel> favoritePrductTitleRequest() async {
     String token = await loadToken();
-    promoDescriptionsUrl += "?api_token=$token&where[0][0]=name&where[0][1]==&where[0][2]=Любимый продукт";
-    final response = await http.get(Uri.parse(promoDescriptionsUrl), headers: {"Accept": "application/json"});
+    promoDescriptionsUrl +=
+        "?api_token=$token&where[0][0]=name&where[0][1]==&where[0][2]=Любимый продукт";
+    final response = await http.get(Uri.parse(promoDescriptionsUrl),
+        headers: {"Accept": "application/json"});
     log('favoritePrductTitleRequest:: get ' + promoDescriptionsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _promoDescriptionsJson = json.decode(response.body);
+      final Map<String, dynamic> _promoDescriptionsJson =
+          json.decode(response.body);
       return FavoriteProductTitleModel.fromJson(_promoDescriptionsJson);
     } else {
-      throw Exception("error fetching promo descriptions. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error fetching promo descriptions. StatusCode: ${response.statusCode}");
     }
   }
 }
@@ -550,7 +609,8 @@ class SmartContactsProvider {
 class GeocodingProvider {
   String geocodingUrl;
 
-  Future<GeocodingModel> getGeocodingReverseResponse({@required Point latLng}) async {
+  Future<GeocodingModel> getGeocodingReverseResponse(
+      {@required Point latLng}) async {
     String token = await loadToken();
     geocodingUrl =
         "$apiHead/clients/api/geocode/reverse?api_token=$token&longitude=${latLng.longitude}&latitude=${latLng.latitude}";
@@ -560,20 +620,25 @@ class GeocodingProvider {
     );
     log('getGeocodingReverseResponse:: get ' + geocodingUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _geocodingModelJson = json.decode(response.body);
+      final Map<String, dynamic> _geocodingModelJson =
+          json.decode(response.body);
       return GeocodingModel.fromJson(_geocodingModelJson);
     } else {
       throw Exception("Geocoding Error");
     }
   }
 
-  Future<GeocodingModel> getGeocodingResponse({@required String address}) async {
+  Future<GeocodingModel> getGeocodingResponse(
+      {@required String address}) async {
     String token = await loadToken();
-    geocodingUrl = "$apiHead/clients/api/geocode?api_token=$token&address=$address";
-    final response = await http.get(Uri.parse(geocodingUrl), headers: {"Accept": "application/json"});
+    geocodingUrl =
+        "$apiHead/clients/api/geocode?api_token=$token&address=$address";
+    final response = await http
+        .get(Uri.parse(geocodingUrl), headers: {"Accept": "application/json"});
     log('getGeocodingResponse:: get ' + geocodingUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> _geocodingModelJson = json.decode(response.body);
+      final Map<String, dynamic> _geocodingModelJson =
+          json.decode(response.body);
       return GeocodingModel.fromJson(_geocodingModelJson);
     } else {
       throw Exception("Geocoding Error");
@@ -585,20 +650,24 @@ class GeocodingProvider {
 class DiverseFoodProvider {
   String _diverseFoodUrl = "$apiHead/clients/api/profile/promo-diverse-food-";
 
-  Future<DiverseFoodPersentListModel> getImInShopPersentListModelResponse() async {
+  Future<DiverseFoodPersentListModel>
+      getImInShopPersentListModelResponse() async {
     String token = await loadToken();
 
-    String diverseFoodUrl = "$apiHead/clients/api/promo-diverse-food-settings?api_token=$token";
+    String diverseFoodUrl =
+        "$apiHead/clients/api/promo-diverse-food-settings?api_token=$token";
     final response = await http.get(
       Uri.parse(diverseFoodUrl),
       headers: {"Accept": "application/json"},
     );
     print('👉 url:: get ' + diverseFoodUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> diverseFoodPersentListJson = json.decode(response.body);
+      final Map<String, dynamic> diverseFoodPersentListJson =
+          json.decode(response.body);
       return DiverseFoodPersentListModel.fromJson(diverseFoodPersentListJson);
     } else {
-      throw Exception("error fetching diverse food persent list. statusCode: ${response.statusCode}");
+      throw Exception(
+          "error fetching diverse food persent list. statusCode: ${response.statusCode}");
     }
   }
 
@@ -606,21 +675,24 @@ class DiverseFoodProvider {
     String token = await loadToken();
     String currentMonth =
         "${DateTime.now().year}-${DateTime.now().month > 9 ? DateTime.now().month : "0${DateTime.now().month}"}";
-    _diverseFoodUrl += "stats?api_token=$token&where[0][0]=month&where[0][1]==&where[0][2]=$currentMonth";
+    _diverseFoodUrl +=
+        "stats?api_token=$token&where[0][0]=month&where[0][1]==&where[0][2]=$currentMonth";
     final response = await http.get(
       Uri.parse(_diverseFoodUrl),
       headers: {"Accept": "application/json"},
     );
     log('diverseFoodStatsResponse:: get ' + _diverseFoodUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> diverseFoodStatJson = json.decode(response.body);
+      final Map<String, dynamic> diverseFoodStatJson =
+          json.decode(response.body);
       return DiverseFoodStatsModel.fromJson(diverseFoodStatJson);
     } else {
       throw Exception("error fetching diverse food stat");
     }
   }
 
-  Future<DiverseFoodPresentDiscountModel> diverseFoodPresentDiscountResponse() async {
+  Future<DiverseFoodPresentDiscountModel>
+      diverseFoodPresentDiscountResponse() async {
     String token = await loadToken();
     _diverseFoodUrl += "discounts?api_token=$token&where[0][0]=start_at&" +
         "where[0][1]=<&where[0][2]=now()&where[1][0]=end_at&where[1][1]=>&where[1][2]=now()";
@@ -630,14 +702,18 @@ class DiverseFoodProvider {
     );
     log('diverseFoodPresentDiscountResponse:: get ' + _diverseFoodUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> diverseFoodPresentDiscountJson = json.decode(response.body);
-      return DiverseFoodPresentDiscountModel.fromJson(diverseFoodPresentDiscountJson);
+      final Map<String, dynamic> diverseFoodPresentDiscountJson =
+          json.decode(response.body);
+      return DiverseFoodPresentDiscountModel.fromJson(
+          diverseFoodPresentDiscountJson);
     } else {
-      throw Exception("error fetching diverse food present discount. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error fetching diverse food present discount. StatusCode: ${response.statusCode}");
     }
   }
 
-  Future<DiverseFoodFutureDiscountModel> diverseFoodFutureDiscountResponse() async {
+  Future<DiverseFoodFutureDiscountModel>
+      diverseFoodFutureDiscountResponse() async {
     String token = await loadToken();
     _diverseFoodUrl += "settings/future-level?api_token=$token";
     final response = await http.get(
@@ -646,17 +722,21 @@ class DiverseFoodProvider {
     );
     log('diverseFoodFutureDiscountResponse:: get ' + _diverseFoodUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> diverseFoodFutureDiscountJson = json.decode(response.body);
-      return DiverseFoodFutureDiscountModel.fromJson(diverseFoodFutureDiscountJson);
+      final Map<String, dynamic> diverseFoodFutureDiscountJson =
+          json.decode(response.body);
+      return DiverseFoodFutureDiscountModel.fromJson(
+          diverseFoodFutureDiscountJson);
     } else {
-      throw Exception("error fetching diverse food present discount. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error fetching diverse food present discount. StatusCode: ${response.statusCode}");
     }
   }
 
   Future<DiverseFoodAssortmentListModel> diverseFoodAssortmentListResponse(
       {@required bool isRated, @required int page}) async {
     String token = await loadToken();
-    _diverseFoodUrl = "$apiHead/clients/api/profile/purchases-month?api_token=$token&page=$page";
+    _diverseFoodUrl =
+        "$apiHead/clients/api/profile/purchases-month?api_token=$token&page=$page";
     if (isRated) {
       _diverseFoodUrl += "&where[0][0]=is_rated&where[0][1]==&where[0][2]=1";
     }
@@ -669,10 +749,13 @@ class DiverseFoodProvider {
     );
     log('diverseFoodAssortmentListResponse:: get ' + _diverseFoodUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> diverseFoodAssortmentListJson = json.decode(response.body);
-      return DiverseFoodAssortmentListModel.fromJson(diverseFoodAssortmentListJson);
+      final Map<String, dynamic> diverseFoodAssortmentListJson =
+          json.decode(response.body);
+      return DiverseFoodAssortmentListModel.fromJson(
+          diverseFoodAssortmentListJson);
     } else {
-      throw Exception("error to fetch Diverse Food Assortment List. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error to fetch Diverse Food Assortment List. StatusCode: ${response.statusCode}");
     }
   }
 }
@@ -681,35 +764,42 @@ class DiverseFoodProvider {
 class FavoriteProductProvider {
   String nowTime =
       "${DateTime.now().year}-${DateTime.now().month < 10 ? "0${DateTime.now().month}" : DateTime.now().month}-${DateTime.now().day < 10 ? "0${DateTime.now().day}" : DateTime.now().day} ${DateTime.now().hour < 10 ? "0${DateTime.now().hour}" : DateTime.now().hour}:${DateTime.now().minute < 10 ? "0${DateTime.now().minute}" : DateTime.now().minute}:${DateTime.now().second < 10 ? "0${DateTime.now().second}" : DateTime.now().second}";
-  String favoriteProductUrl = "$apiHead/clients/api/profile/active-promo-favorite-assortments";
+  String favoriteProductUrl =
+      "$apiHead/clients/api/profile/active-promo-favorite-assortments";
 
   // get favorite product
   Future<FavoriteProductModel> getFavoriteProductResponse() async {
     String token = await loadToken();
     String shopUuid = await loadShopUuid();
     favoriteProductUrl += "?api_token=$token&per_page=20&store_uuid=$shopUuid";
-    favoriteProductUrl += "&where[0][0]=active_to&where[0][1]=>&where[0][2]=$nowTime";
+    favoriteProductUrl +=
+        "&where[0][0]=active_to&where[0][1]=>&where[0][2]=$nowTime";
     final response = await http.get(
       Uri.parse(favoriteProductUrl),
       headers: {"Accept": "application/json"},
     );
     print('🤍:: getFavoriteProductResponse get ' + favoriteProductUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> favoriteProductJson = json.decode(response.body);
+      final Map<String, dynamic> favoriteProductJson =
+          json.decode(response.body);
       return FavoriteProductModel.fromJson(favoriteProductJson);
     } else {
-      throw Exception("error fetching favorite Product. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "error fetching favorite Product. StatusCode: ${response.statusCode}");
     }
   }
 
   //set favorite product
 
-  Future<String> setFavoriteProductResponse({@required String assortmentUuid, @required String variantUuid}) async {
+  Future<String> setFavoriteProductResponse(
+      {@required String assortmentUuid, @required String variantUuid}) async {
     String token = await loadToken();
 
-    favoriteProductUrl = "$apiHead/clients/api/profile/promo-favorite-assortment-variants/$variantUuid/activate";
+    favoriteProductUrl =
+        "$apiHead/clients/api/profile/promo-favorite-assortment-variants/$variantUuid/activate";
     final response = await http.post(Uri.parse(favoriteProductUrl),
-        headers: {"Accept": "application/json"}, body: {"api_token": token, "assortment_uuid": assortmentUuid});
+        headers: {"Accept": "application/json"},
+        body: {"api_token": token, "assortment_uuid": assortmentUuid});
     print('🤍:: setFavoriteProductResponse post ' + favoriteProductUrl);
     // print('api_token:: ' + token);
     // print('assortment_uuid:: ' + assortmentUuid);
@@ -728,20 +818,27 @@ class FavoriteProductProvider {
   }
 
   //get variant uuid
-  Future<FavoriteProductVariantUuidModel> getFavoriteProductVariantUuidResponse() async {
+  Future<FavoriteProductVariantUuidModel>
+      getFavoriteProductVariantUuidResponse() async {
     String token = await loadToken();
-    favoriteProductUrl = "$apiHead/clients/api/profile/promo-favorite-assortment-variants?api_token=$token";
-    favoriteProductUrl += "&where[0][0]=can_be_activated_till&where[0][1]=>&where[0][2]=$nowTime";
+    favoriteProductUrl =
+        "$apiHead/clients/api/profile/promo-favorite-assortment-variants?api_token=$token";
+    favoriteProductUrl +=
+        "&where[0][0]=can_be_activated_till&where[0][1]=>&where[0][2]=$nowTime";
     final response = await http.get(
       Uri.parse(favoriteProductUrl),
       headers: {"Accept": "application/json"},
     );
-    print('🤍🤍🤍🤍🤍🤍🤍:: getFavoriteProductVariantUuidResponse get ' + favoriteProductUrl);
+    print('🤍🤍🤍🤍🤍🤍🤍:: getFavoriteProductVariantUuidResponse get ' +
+        favoriteProductUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> favoriteProductVariantUuidJson = json.decode(response.body);
-      return FavoriteProductVariantUuidModel.fromJson(favoriteProductVariantUuidJson);
+      final Map<String, dynamic> favoriteProductVariantUuidJson =
+          json.decode(response.body);
+      return FavoriteProductVariantUuidModel.fromJson(
+          favoriteProductVariantUuidJson);
     } else {
-      throw Exception("error fetching variant uuid. StatusCode: ${response.statusCode} ");
+      throw Exception(
+          "error fetching variant uuid. StatusCode: ${response.statusCode} ");
     }
   }
 }
@@ -750,7 +847,8 @@ class FavoriteProductProvider {
 class RecomendationProvider {
   String recomendationUrl;
 
-  Future<List<AssortmentsListModel>> getRecomendationListResponse({@required int page}) async {
+  Future<List<AssortmentsListModel>> getRecomendationListResponse(
+      {@required int page}) async {
     String token = await loadToken();
     String shopUuid = await loadShopUuid();
 
@@ -766,9 +864,11 @@ class RecomendationProvider {
       Uri.parse(recomendationUrl),
       headers: {"Accept": "application/json"},
     );
-    log('get рекомендуемые товары getRecomendationListResponse:: ' + recomendationUrl);
+    log('get рекомендуемые товары getRecomendationListResponse:: ' +
+        recomendationUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> recomendationListJson = json.decode(response.body);
+      final Map<String, dynamic> recomendationListJson =
+          json.decode(response.body);
       return RecomendationListModel.fromJson(recomendationListJson).data;
     } else {
       throw Exception("error fetching Recomendation list");
@@ -780,7 +880,8 @@ class RecomendationProvider {
 class NotificationsProvider {
   String notificationsUrl = "$apiHead/clients/api/profile/notifications";
 
-  Future<NotificationsListModel> getNotificationListResponse({@required int page}) async {
+  Future<NotificationsListModel> getNotificationListResponse(
+      {@required int page}) async {
     String token = await loadToken();
     final response = await http.get(
       Uri.parse(notificationsUrl + "?api_token=$token&page=$page"),
@@ -788,7 +889,8 @@ class NotificationsProvider {
     );
     log('url:: get ' + notificationsUrl + "?api_token=$token&page=$page");
     if (response.statusCode == 200) {
-      final Map<String, dynamic> notificationListJson = json.decode(response.body);
+      final Map<String, dynamic> notificationListJson =
+          json.decode(response.body);
       return NotificationsListModel.fromJson(notificationListJson);
     } else {
       throw Exception('error fetching notification list!!!');
@@ -811,10 +913,15 @@ class OrderProvider {
     String urlOrderMark = "$orderUrl/products/$orderProductUuid/set-rating";
     final response = await http.post(
       Uri.parse(urlOrderMark),
-      body: {"value": value.toInt().toString(), "comment": comment, "api_token": token},
+      body: {
+        "value": value.toInt().toString(),
+        "comment": comment,
+        "api_token": token
+      },
       headers: {"Accept": "application/json"},
     );
-    print('🟢:: post setRatingForAssortmentInOrderResponse:::::::: ' + urlOrderMark);
+    print('🟢:: post setRatingForAssortmentInOrderResponse:::::::: ' +
+        urlOrderMark);
     // print('value:: ' + value.toInt().toString() + ', comment:: ' + comment + ', api_token:: ' + token);
     if (response.statusCode == 204) {
       return true;
@@ -835,7 +942,8 @@ class OrderProvider {
     });
     print('🟢:: get getOrderListResponse::::::::: ' + orderUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> orderListResponseJson = json.decode(response.body);
+      final Map<String, dynamic> orderListResponseJson =
+          json.decode(response.body);
       return OrderListModel.fromJson(orderListResponseJson);
     } else {
       throw Exception("Server's error: ${response.statusCode}");
@@ -849,11 +957,14 @@ class OrderProvider {
     token = await loadToken();
     orderUrl = "$orderUrl/$orderId?api_token=$token";
 
-    final response = await http.get(Uri.parse(orderUrl), headers: {'Accept': 'application/json'});
+    final response = await http
+        .get(Uri.parse(orderUrl), headers: {'Accept': 'application/json'});
     print('url 🛒🟢 orderDetailseResponse get:: ' + orderUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> orderDetailsResponseJson = json.decode(response.body);
-      return OrderDetailsAndCalculateResponseModel.fromJson(orderDetailsResponseJson);
+      final Map<String, dynamic> orderDetailsResponseJson =
+          json.decode(response.body);
+      return OrderDetailsAndCalculateResponseModel.fromJson(
+          orderDetailsResponseJson);
     } else {
       throw Exception(response.body.toString());
     }
@@ -861,21 +972,21 @@ class OrderProvider {
 
   // Создать заказ
   Future<OrderCreateResponseModel> createOrderResponse({
-    @required String clientCreditCardUuid,
-    String clientComment,
-    String promocode,
-    String clientEmail,
-    String address,
-    int floor,
-    int entrance,
-    int apartmentNumber,
-    String intercomCode,
-    String plannedDeliveryDatetimeFrom,
-    String plannedDeliveryDatetimeTo,
-    String orderDeliveryTypeId,
-    String orderPaymentTypeId,
-    int subtractBonusesCount,
-    List<ProductModelForOrderRequest> productModelForOrderRequestList,
+    String? clientCreditCardUuid,
+    String? clientComment,
+    String? promocode,
+    String? clientEmail,
+    String? address,
+    int? floor,
+    int? entrance,
+    int? apartmentNumber,
+    String? intercomCode,
+    String? plannedDeliveryDatetimeFrom,
+    String? plannedDeliveryDatetimeTo,
+    String? orderDeliveryTypeId,
+    String? orderPaymentTypeId,
+    int? subtractBonusesCount,
+    List<ProductModelForOrderRequest>? productModelForOrderRequestList,
   }) async {
     token = await loadToken();
     String storeUuid = await loadShopUuid();
@@ -903,14 +1014,22 @@ class OrderProvider {
           },
         "promocode": "test15"
       },
-      "products": List<dynamic>.from(productModelForOrderRequestList.map((x) => x.toJson()))
+      "products": List<dynamic>.from(
+          productModelForOrderRequestList?.map((x) => x.toJson()) ?? [])
     });
     final response = await http.post(Uri.parse("$orderUrl"),
-        body: body, headers: {'Accept': 'application/json', 'Content-Type': 'application/json'});
+        body: body,
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json'
+        });
     print('url 🛒🟢 createOrderResponse post:: ' + orderUrl + "а body ниже:");
-    print('body:: ' + orderDeliveryTypeId + 'есть еще другие параметры см сервисы');
+    print('body:: ' +
+        (orderDeliveryTypeId ?? '') +
+        'есть еще другие параметры см сервисы');
     if (response.statusCode == 201) {
-      final Map<String, dynamic> orderCreatedResponseJson = json.decode(response.body);
+      final Map<String, dynamic> orderCreatedResponseJson =
+          json.decode(response.body);
       return OrderCreateResponseModel.fromJson(orderCreatedResponseJson);
     } else {
       if (response.statusCode == 400) {
@@ -927,11 +1046,11 @@ class OrderProvider {
 
   //services Расчёт заказа
   Future<OrderDetailsAndCalculateResponseModel> orderCalculateResponse({
-    @required String orderDeliveryTypeId,
-    @required String orderPaymentTypeId,
-    int subtractBonusesCount,
-    String promocode,
-    @required List<ProductModelForOrderRequest> productModelForOrderRequestList,
+    String? orderDeliveryTypeId,
+    required String? orderPaymentTypeId,
+    int? subtractBonusesCount,
+    String? promocode,
+    required List<ProductModelForOrderRequest>? productModelForOrderRequestList,
   }) async {
     token = await loadToken();
     String storeUuid = await loadShopUuid();
@@ -941,24 +1060,32 @@ class OrderProvider {
         if (promocode != null && promocode != "") "promocode": promocode,
         "order_payment_type_id": orderPaymentTypeId,
         "store_user_uuid": storeUuid,
-        "order_delivery_type_id": orderDeliveryTypeId == null ? "delivery" : orderDeliveryTypeId,
-        if (subtractBonusesCount != null && subtractBonusesCount != 0) "paid_bonus": subtractBonusesCount,
+        "order_delivery_type_id":
+            orderDeliveryTypeId == null ? "delivery" : orderDeliveryTypeId,
+        if (subtractBonusesCount != null && subtractBonusesCount != 0)
+          "paid_bonus": subtractBonusesCount,
       },
-      "products": List<dynamic>.from(productModelForOrderRequestList.map((x) => x.toJson()))
+      "products": List<dynamic>.from(
+          productModelForOrderRequestList.map((x) => x.toJson()))
     });
     String url = "$orderUrl/calculate";
-    final response = await http.post(Uri.parse(url),
-        body: (body), headers: {'Accept': 'application/json', 'Content-Type': 'application/json'});
+    final response = await http.post(Uri.parse(url), body: (body), headers: {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    });
     log('--url 🛒📖orderCalculateResponse post 🟡:: ' + url);
     print('--body 🛒📖:: ' + body);
     print('--response 🛒📖:: ' + response.body);
 
     if (response.statusCode == 200) {
-      final Map<String, dynamic> orderCalculateResponseJson = json.decode(response.body);
+      final Map<String, dynamic> orderCalculateResponseJson =
+          json.decode(response.body);
       final dynamic totalPriceForProductsWithDiscount =
-          orderCalculateResponseJson['data']['total_price_for_products_with_discount'];
+          orderCalculateResponseJson['data']
+              ['total_price_for_products_with_discount'];
       if (orderCalculateResponseJson['data']["promocode"] != null)
-        log('--response 🛒📖🎁 promocode:: ' + orderCalculateResponseJson['data']["promocode"]);
+        log('--response 🛒📖🎁 promocode:: ' +
+            orderCalculateResponseJson['data']["promocode"]);
       if (totalPriceForProductsWithDiscount != null) {
         double totalPriceAsDouble;
         if (totalPriceForProductsWithDiscount is int) {
@@ -970,10 +1097,12 @@ class OrderProvider {
         }
 
         SharedPreferences prefs = await SharedPreferences.getInstance();
-        await prefs.setDouble(SharedKeys.totalPriceForProductsWithDiscount, totalPriceAsDouble);
+        await prefs.setDouble(
+            SharedKeys.totalPriceForProductsWithDiscount, totalPriceAsDouble);
       }
 
-      return OrderDetailsAndCalculateResponseModel.fromJson(orderCalculateResponseJson);
+      return OrderDetailsAndCalculateResponseModel.fromJson(
+          orderCalculateResponseJson);
     } else {
       // print(toCurl(response.request));
       throw Exception(response.body.toString());
@@ -989,8 +1118,8 @@ class BasketProvider {
   Future<bool> fillBasketFromORderResponse(String orderUuid) async {
     token = await loadToken();
     basketUrl += "/fill-from-order/$orderUuid";
-    final response =
-        await http.post(Uri.parse(basketUrl), headers: {"Accept": "application/json"}, body: {"api_token": token});
+    final response = await http.post(Uri.parse(basketUrl),
+        headers: {"Accept": "application/json"}, body: {"api_token": token});
     log('url 🛒  fillBasketFromORderResponse post:: ' + basketUrl);
     // print('api_token:: ' + token);
     if (response.statusCode == 204) {
@@ -1026,10 +1155,12 @@ class BasketProvider {
   //Заполнение корзины товарами из списка покупок
   Future<bool> addShoppingListToBasket(String shoppingListUuid) async {
     token = await loadToken();
-    basketUrl = "$apiHead/clients/api/profile/shopping-cart/fill-from-shopping-list/{$shoppingListUuid}";
-    final response =
-        await http.post(Uri.parse(basketUrl), headers: {"Accept": "application/json"}, body: {"api_token": token});
-    log('url 🛒 Заполнение корзины товарами из списка покупок post:: ' + basketUrl);
+    basketUrl =
+        "$apiHead/clients/api/profile/shopping-cart/fill-from-shopping-list/{$shoppingListUuid}";
+    final response = await http.post(Uri.parse(basketUrl),
+        headers: {"Accept": "application/json"}, body: {"api_token": token});
+    log('url 🛒 Заполнение корзины товарами из списка покупок post:: ' +
+        basketUrl);
     // print('api_token:: ' + token);
     if (response.statusCode == 204) {
       return true;
@@ -1041,9 +1172,13 @@ class BasketProvider {
   Future<bool> addProductInBasket(String productUuid, double quantity) async {
     token = await loadToken();
     basketUrl += "/assortments";
-    final response = await http.post(Uri.parse(basketUrl),
-        body: {'uuid': productUuid, if (quantity != null) 'quantity': quantity.toStringAsFixed(2), 'api_token': token},
-        headers: {'Accept': 'application/json'});
+    final response = await http.post(Uri.parse(basketUrl), body: {
+      'uuid': productUuid,
+      if (quantity != null) 'quantity': quantity.toStringAsFixed(2),
+      'api_token': token
+    }, headers: {
+      'Accept': 'application/json'
+    });
     log('url 🛒addProductInBasket post:: ' + basketUrl);
     // print('uuid:: ' + productUuid);
     // if (quantity != null) print('quantity:: ' + quantity.toStringAsFixed(2));
@@ -1057,9 +1192,12 @@ class BasketProvider {
 
   Future<bool> reomoveProductFromBasket(String productUuid) async {
     token = await loadToken();
-    final response = await http.delete(Uri.parse(basketUrl + "/assortments/$productUuid?api_token=$token"),
+    final response = await http.delete(
+        Uri.parse(basketUrl + "/assortments/$productUuid?api_token=$token"),
         headers: {'Accept': 'application/json'});
-    log('url 🛒reomoveProductFromBasket delete:: ' + basketUrl + "/assortments/$productUuid?api_token=$token");
+    log('url 🛒reomoveProductFromBasket delete:: ' +
+        basketUrl +
+        "/assortments/$productUuid?api_token=$token");
     if (response.statusCode == 204) {
       return true;
     } else {
@@ -1067,13 +1205,24 @@ class BasketProvider {
     }
   }
 
-  Future<bool> updateProductInBasket({String productUuid, double quantity}) async {
+  Future<bool> updateProductInBasket(
+      {String productUuid, double quantity}) async {
     token = await loadToken();
-    final response = await http.post(Uri.parse(basketUrl + "/assortments/update"),
-        body: {'uuid': productUuid.toString(), 'quantity': quantity.toStringAsFixed(2), 'api_token': token},
-        headers: {'Accept': 'application/json'});
-    log('url 🛒updateProductInBasket post:: ' + basketUrl + "/assortments/update");
-    print('body: uuid:: ' + productUuid.toString() + ', quantity:: ' + quantity.toStringAsFixed(2));
+    final response =
+        await http.post(Uri.parse(basketUrl + "/assortments/update"), body: {
+      'uuid': productUuid.toString(),
+      'quantity': quantity.toStringAsFixed(2),
+      'api_token': token
+    }, headers: {
+      'Accept': 'application/json'
+    });
+    log('url 🛒updateProductInBasket post:: ' +
+        basketUrl +
+        "/assortments/update");
+    print('body: uuid:: ' +
+        productUuid.toString() +
+        ', quantity:: ' +
+        quantity.toStringAsFixed(2));
     if (response.statusCode == 204) {
       return true;
     } else {
@@ -1083,9 +1232,11 @@ class BasketProvider {
 
   Future<bool> removeAllBasket() async {
     token = await loadToken();
-    final response = await http
-        .delete(Uri.parse("$basketUrl/assortments?api_token=$token"), headers: {'Accept': 'application/json'});
-    log('url 🛒removeAllBasket delete:: ' + "$basketUrl/assortments?api_token=$token");
+    final response = await http.delete(
+        Uri.parse("$basketUrl/assortments?api_token=$token"),
+        headers: {'Accept': 'application/json'});
+    log('url 🛒removeAllBasket delete:: ' +
+        "$basketUrl/assortments?api_token=$token");
     if (response.statusCode == 204) {
       return true;
     } else {
@@ -1136,7 +1287,8 @@ class AssrtmentCommentsProvider {
     );
     log('getAssortmentCommentsResponse:: get ' + url);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> assortmentCommentsJson = json.decode(response.body);
+      final Map<String, dynamic> assortmentCommentsJson =
+          json.decode(response.body);
       return AssortmentCommentsModel.fromJson(assortmentCommentsJson);
     } else {
       throw Exception('Error fetching assortment comments!!!');
@@ -1146,10 +1298,13 @@ class AssrtmentCommentsProvider {
 
 class GetCodeByPhoneProvider {
   Future<bool> getAuthResponse({@required String phone}) async {
-    final response = await http.post(Uri.parse("$apiHead/clients/api/auth/login-via-phone"),
-        body: {"phone": phone}, headers: {"Accept": "application/json"});
+    final response = await http.post(
+        Uri.parse("$apiHead/clients/api/auth/login-via-phone"),
+        body: {"phone": phone},
+        headers: {"Accept": "application/json"});
     // print(toCurl(response.request));
-    log('$phone🔑 getAuthResponse post ' + "$apiHead/clients/api/auth/login-via-phone");
+    log('$phone🔑 getAuthResponse post ' +
+        "$apiHead/clients/api/auth/login-via-phone");
     // print('phone:: ' + phone);
     if (response.statusCode == 201) {
       return true;
@@ -1160,10 +1315,14 @@ class GetCodeByPhoneProvider {
 }
 
 class GetTokenByPhoneProvider {
-  Future<String> getTokenResponse({@required String phone, @required String code}) async {
-    final response = await http.post(Uri.parse("$apiHead/clients/api/auth/login-via-phone"),
-        body: {'phone': phone, 'code': code}, headers: {'Accept': 'application/json'});
-    print('🔑 getTokenResponse post ' + "$apiHead/clients/api/auth/login-via-phone");
+  Future<String> getTokenResponse(
+      {@required String phone, @required String code}) async {
+    final response = await http.post(
+        Uri.parse("$apiHead/clients/api/auth/login-via-phone"),
+        body: {'phone': phone, 'code': code},
+        headers: {'Accept': 'application/json'});
+    print('🔑 getTokenResponse post ' +
+        "$apiHead/clients/api/auth/login-via-phone");
     // print('phone:: ' + phone + ', code::' + code);
     if (response.statusCode == 200) {
       final Map<String, dynamic> tokenJson = json.decode(response.body);
@@ -1180,7 +1339,8 @@ class GetTokenByPhoneProvider {
 
 //stores class
 class ShopsListProvider {
-  Future<AddressesShopModel> getNearbyStoreModel({@required Point position}) async {
+  Future<AddressesShopModel> getNearbyStoreModel(
+      {@required Point position}) async {
     String token = await loadToken();
     String url = '$apiHead/clients/api/stores/find-nearby';
     if (token != null) {
@@ -1188,11 +1348,15 @@ class ShopsListProvider {
     }
     // url += "&latitude=${position.latitude}&longitude=${position.longitude}&limit=1";
     url += "&latitude=${position.latitude}&longitude=${position.longitude}";
-    var response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
-    print('url:: 📍 магазины рядом get Future<FindNearbyStoreModel> getNearbyStoreModel ' + url);
+    var response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    print(
+        'url:: 📍 магазины рядом get Future<FindNearbyStoreModel> getNearbyStoreModel ' +
+            url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> jsonResponse = json.decode(response.body);
-      if (jsonResponse.containsKey("data") && (jsonResponse["data"] as List).isNotEmpty) {
+      if (jsonResponse.containsKey("data") &&
+          (jsonResponse["data"] as List).isNotEmpty) {
         final nearestShopJson = jsonResponse["data"][0];
         print('📍 Найден ближайший магазин: ${nearestShopJson["uuid"]}');
         return AddressesShopModel.fromJson(nearestShopJson);
@@ -1206,9 +1370,15 @@ class ShopsListProvider {
   }
 
   Future<AddressesShopListModel> getShopsList(
-      {bool hasParking, bool hasReadyMeals, bool hasAtms, bool isfavorite, bool isOpenNow, String searchText}) async {
+      {bool hasParking,
+      bool hasReadyMeals,
+      bool hasAtms,
+      bool isfavorite,
+      bool isOpenNow,
+      String searchText}) async {
     String token = await loadToken();
-    String url = '$apiHead/clients/api/stores?per_page=1000&order_by[address]=asc';
+    String url =
+        '$apiHead/clients/api/stores?per_page=1000&order_by[address]=asc';
     if (token != 'guest') {
       url += '&api_token=$token';
     }
@@ -1228,7 +1398,8 @@ class ShopsListProvider {
       url += "&where[4][0]=is_favorite&where[4][1]=like&where[4][2]=true";
     }
 
-    final response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    final response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
     log('url:: 🏠 адреса магазинов Future<ShopsModel> getShopsList: ' + url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> shopsListJson = json.decode(response.body);
@@ -1248,7 +1419,8 @@ class ShopDetailsProvider {
     if (token != 'guest') {
       url += '?api_token=$token';
     }
-    final response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    final response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
     print('url:: get ' + url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> shopDetailJson = json.decode(response.body);
@@ -1277,12 +1449,14 @@ class CatalogsProvider {
     if (catalogUuid != null) {
       print(
           'if (catalogUuid != null) if ДОЧЕРНИЕ => url = url + &where[1][0]=catalog_uuid&where[1][1]=like&where[1][2]=$catalogUuid');
-      url = url + '&where[1][0]=catalog_uuid&where[1][1]=like&where[1][2]=$catalogUuid';
+      url = url +
+          '&where[1][0]=catalog_uuid&where[1][1]=like&where[1][2]=$catalogUuid';
       if (shopUuid != null) {
         print('if (shopUuid != null) => url = url + "&store_uuid=$shopUuid"');
         url = url + "&store_uuid=$shopUuid";
       }
-      response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+      response = await http
+          .get(Uri.parse(url), headers: {'Accept': 'application/json'});
       if (response.statusCode == 200) {
         final Map<String, dynamic> catalogsJson = json.decode(response.body);
         return CatalogsModel.fromJson(catalogsJson).data;
@@ -1293,28 +1467,38 @@ class CatalogsProvider {
       if (shopUuid != null) {
         print(
             'if (catalogUuid != null) else РОДИТЕЛЬСКИЕ => url = url + &where[1][0]=level&where[1][1]=like&where[1][2]= 1 а потом 2 &store_uuid=$shopUuid');
-        String urlLevel1 = '$url&where[1][0]=level&where[1][1]=like&where[1][2]=1&store_uuid=$shopUuid';
-        String urlLevel2 = '$url&where[1][0]=level&where[1][1]=like&where[1][2]=2&store_uuid=$shopUuid';
+        String urlLevel1 =
+            '$url&where[1][0]=level&where[1][1]=like&where[1][2]=1&store_uuid=$shopUuid';
+        String urlLevel2 =
+            '$url&where[1][0]=level&where[1][1]=like&where[1][2]=2&store_uuid=$shopUuid';
         print('url:: категории urlLevel1️⃣ get: ' + urlLevel1);
-        var responseLevel1 = await http.get(Uri.parse(urlLevel1), headers: {'Accept': 'application/json'});
+        var responseLevel1 = await http
+            .get(Uri.parse(urlLevel1), headers: {'Accept': 'application/json'});
         print('url:: категории urlLevel2️⃣ get: ' + urlLevel2);
-        var responseLevel2 = await http.get(Uri.parse(urlLevel2), headers: {'Accept': 'application/json'});
+        var responseLevel2 = await http
+            .get(Uri.parse(urlLevel2), headers: {'Accept': 'application/json'});
 
-        if (responseLevel1.statusCode == 200 && responseLevel2.statusCode == 200) {
+        if (responseLevel1.statusCode == 200 &&
+            responseLevel2.statusCode == 200) {
           var topLevelJson = json.decode(responseLevel1.body)['data'] as List;
-          var secondLevelJson = json.decode(responseLevel2.body)['data'] as List;
+          var secondLevelJson =
+              json.decode(responseLevel2.body)['data'] as List;
 
-          List<CatalogListModel> topLevelCatalogs = topLevelJson.map((e) => CatalogListModel.fromJson(e)).toList();
+          List<CatalogListModel> topLevelCatalogs =
+              topLevelJson.map((e) => CatalogListModel.fromJson(e)).toList();
           List<CatalogListModel> secondLevelCatalogs =
               secondLevelJson.map((e) => CatalogListModel.fromJson(e)).toList();
 
           Map<String, List<CatalogListModel>> secondLevelMap = {};
           for (var secondLevelCatalog in secondLevelCatalogs) {
-            secondLevelMap.putIfAbsent(secondLevelCatalog.catalogUuid, () => []).add(secondLevelCatalog);
+            secondLevelMap
+                .putIfAbsent(secondLevelCatalog.catalogUuid, () => [])
+                .add(secondLevelCatalog);
           }
 
           for (var topLevelCatalog in topLevelCatalogs) {
-            topLevelCatalog.subcatalog = secondLevelMap[topLevelCatalog.uuid] ?? [];
+            topLevelCatalog.subcatalog =
+                secondLevelMap[topLevelCatalog.uuid] ?? [];
           }
 
           return topLevelCatalogs;
@@ -1354,7 +1538,8 @@ class AssortmentsProvider {
     this.searchText,
   });
 
-  Future<List<AssortmentsListModel>> getAssortmentsForImInShop({@required List<String> assortmentsUuidList}) async {
+  Future<List<AssortmentsListModel>> getAssortmentsForImInShop(
+      {@required List<String> assortmentsUuidList}) async {
     String token = await loadToken();
     String shopUuid = await loadShopUuid();
     String _urlForImInShopList = shopUuid == null || shopUuid == ""
@@ -1369,9 +1554,11 @@ class AssortmentsProvider {
         _urlForImInShopList += "&where[0][2][$i]=${assortmentsUuidList[i]}";
       }
     }
-    final response = await http.get(Uri.parse(_urlForImInShopList), headers: {'Accept': 'application/json'});
-    print('url:: 🍏 get получить продукты (AssortmentsProvider) getAssortmentsForImInShop ::::::::::::::::::::::::: ' +
-        _urlForImInShopList);
+    final response = await http.get(Uri.parse(_urlForImInShopList),
+        headers: {'Accept': 'application/json'});
+    print(
+        'url:: 🍏 get получить продукты (AssortmentsProvider) getAssortmentsForImInShop ::::::::::::::::::::::::: ' +
+            _urlForImInShopList);
     if (response.statusCode == 200) {
       final Map<String, dynamic> assortmentsJson = json.decode(response.body);
       return AssortmentsModel.fromJson(assortmentsJson).data;
@@ -1381,10 +1568,13 @@ class AssortmentsProvider {
   }
 
   Future<List<AssortmentsListModel>> getAssortmentsForPagination(
-      {bool isAllSubcatalogsWithoutFavorite = false, String subcatalogUuid}) async {
-    print('========== !!!!!!!!!! getAssortmentsForPagination subcatalogUuid:::::::::::::::::::::::');
+      {bool isAllSubcatalogsWithoutFavorite = false,
+      String subcatalogUuid}) async {
+    print(
+        '========== !!!!!!!!!! getAssortmentsForPagination subcatalogUuid:::::::::::::::::::::::');
     print('$subcatalogUuid');
-    print('----------------------------------------------------------------------------------------');
+    print(
+        '----------------------------------------------------------------------------------------');
     log('========== !!!!!!!!!! getAssortmentsForPagination subcatalogUuid = $subcatalogUuid');
     String token = await loadToken();
     //=>
@@ -1393,14 +1583,16 @@ class AssortmentsProvider {
         ? '$apiHead/clients/api/assortments?page=$currentPage&per_page=20&order_by[name]=asc'
         : '$apiHead/clients/api/stores/$shopUuid/assortments?page=$currentPage&per_page=20&order_by[name]=asc';
     // if (isAllSubcatalogsWithoutFavorite != null && isAllSubcatalogsWithoutFavorite && (isFavorite == null || isFavorite == false) && (isPromoAssortment == null || isPromoAssortment == false)) {
-    if (isAllSubcatalogsWithoutFavorite != null && isAllSubcatalogsWithoutFavorite) {
+    if (isAllSubcatalogsWithoutFavorite != null &&
+        isAllSubcatalogsWithoutFavorite) {
       // это срабатывает когда нажимаем на подкатегорию, 5 часов мучилась (и их не залогаю) чтобы сделать там норм подгрузку но тогда непонятно почему избранное перестает работать
       // и поэтому я вписала per_page=500, но когда открываем категорию 2го уровня у которой нет подкатегорий 3го уровня, там все идеально подгружается
       // например колбасные изделия
       _url =
           "$apiHead/clients/api/stores/$shopUuid/assortments?page=1&order_by[catalog_name]=asc&per_page=500&api_token=$token&where[1][0]=catalog_uuid&where[1][1]=in&store_uuid=$shopUuid&where[1][2][0]=$subcatalogUuid";
       if (isFavorite != null && isFavorite == true) {
-        _url = _url + '&where[2][0]=is_favorite&where[2][1]=like&where[2][2]=true';
+        _url =
+            _url + '&where[2][0]=is_favorite&where[2][1]=like&where[2][2]=true';
       }
       if (isPromoAssortment != null && isPromoAssortment != false) {
         _url += "&where[6][0]=has_yellow_price&where[6][1]==&where[6][2]=true";
@@ -1408,16 +1600,20 @@ class AssortmentsProvider {
     } else {
       if (subcatalogUuid != null) {
         log('subcatalogUuid = $subcatalogUuid');
-        _url = _url + '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$subcatalogUuid';
+        _url = _url +
+            '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$subcatalogUuid';
       }
       if (catalogUuid != null) {
-        _url = _url + '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$catalogUuid';
+        _url = _url +
+            '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$catalogUuid';
       }
       if (searchText != null) {
-        _url = _url + '&where[1][0]=name&where[1][1]=ilike&where[1][2]=%$searchText%';
+        _url = _url +
+            '&where[1][0]=name&where[1][1]=ilike&where[1][2]=%$searchText%';
       }
       if (isFavorite != null && isFavorite == true) {
-        _url = _url + '&where[2][0]=is_favorite&where[2][1]=like&where[2][2]=true';
+        _url =
+            _url + '&where[2][0]=is_favorite&where[2][1]=like&where[2][2]=true';
       }
       if (isRecommendations != null) {
         _url += "&where[3][0]=properties&where[3][1]=in&where[3][2][0]=%D0%A0%D0%B5%D0%BA%D0%BE%D0%BC%D" +
@@ -1430,7 +1626,8 @@ class AssortmentsProvider {
         }
       }
       if (uuidForAllProductsInCatalog != null) {
-        _url += "&where[5][0]=catalog_with_children_uuid&where[5][1]==&where[5][2]=$uuidForAllProductsInCatalog";
+        _url +=
+            "&where[5][0]=catalog_with_children_uuid&where[5][1]==&where[5][2]=$uuidForAllProductsInCatalog";
       }
       //=>
       if (isPromoAssortment != null && isPromoAssortment != false) {
@@ -1439,7 +1636,8 @@ class AssortmentsProvider {
 
       if (activeTagsList != null && activeTagsList.isNotEmpty) {
         for (int i = 0; i < activeTagsList.length; i++) {
-          _url = _url + '&where[${i + 7}][0]=tags&where[${i + 7}][1]=like&where[${i + 7}][2]=${activeTagsList[i]}';
+          _url = _url +
+              '&where[${i + 7}][0]=tags&where[${i + 7}][1]=like&where[${i + 7}][2]=${activeTagsList[i]}';
         }
       }
       if (token != "guest") {
@@ -1449,7 +1647,8 @@ class AssortmentsProvider {
       //   url = url + "&store_uuid=$shopUuid";
       // }
     }
-    final response = await http.get(Uri.parse(_url), headers: {'Accept': 'application/json'});
+    final response = await http
+        .get(Uri.parse(_url), headers: {'Accept': 'application/json'});
     print(
         '-url:: get получить продукты 🍎 (AssortmentsProvider) getAssortmentsForPagination ::::::::::::::::::::::::: ' +
             _url);
@@ -1468,10 +1667,12 @@ class AssortmentsProvider {
         ? '$apiHead/clients/api/assortments?page=$currentPage'
         : '$apiHead/clients/api/stores/{$shopUuid}/assortments?page=$currentPage';
     if (catalogUuid != null) {
-      url = url + '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$catalogUuid';
+      url = url +
+          '&where[0][0]=catalog_uuid&where[0][1]=like&where[0][2]=$catalogUuid';
     }
     if (searchText != null) {
-      url = url + '&where[1][0]=name&where[1][1]=ilike&where[1][2]=%$searchText%';
+      url =
+          url + '&where[1][0]=name&where[1][1]=ilike&where[1][2]=%$searchText%';
     }
 
     if (isFavorite != null) {
@@ -1479,11 +1680,13 @@ class AssortmentsProvider {
     }
 
     if (brandName != null) {
-      url += "&where[3][0]=assortment_brand_name&where[3][1]=ilike&where[3][2]=%$brandName%";
+      url +=
+          "&where[3][0]=assortment_brand_name&where[3][1]=ilike&where[3][2]=%$brandName%";
     }
     if (activeTagsList != null && activeTagsList.isNotEmpty) {
       for (int i = 0; i < activeTagsList.length; i++) {
-        url = url + '&where[${i + 4}][0]=tags&where[${i + 4}][1]=like&where[${i + 4}][2]=${activeTagsList[i]}';
+        url = url +
+            '&where[${i + 4}][0]=tags&where[${i + 4}][1]=like&where[${i + 4}][2]=${activeTagsList[i]}';
       }
     }
     if (barcodes != null) {
@@ -1501,8 +1704,11 @@ class AssortmentsProvider {
       url = url + "&store_uuid=$shopUuid";
     }
 
-    final response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
-    print('url:: get 🍎 получить продукты (AssortmentsProvider) getAssortments ::::::::::::::::::::::::: ' + url);
+    final response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    print(
+        'url:: get 🍎 получить продукты (AssortmentsProvider) getAssortments ::::::::::::::::::::::::: ' +
+            url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> assortmentsJson = json.decode(response.body);
       return AssortmentsModel.fromJson(assortmentsJson);
@@ -1520,7 +1726,8 @@ class TagsProvider {
       tagsUrl +=
           "?api_token=$token&page=1&per_page=1000&order_by[name]=asc&where[0][0]=fixed_in_filters&where[0][1]=in&where[0][2]=true";
     }
-    final response = await http.get(Uri.parse(tagsUrl), headers: {'Accept': 'application/json'});
+    final response = await http
+        .get(Uri.parse(tagsUrl), headers: {'Accept': 'application/json'});
     print('url:: get ' + tagsUrl);
     if (response.statusCode == 200) {
       final Map<String, dynamic> tagsJson = json.decode(response.body);
@@ -1563,10 +1770,12 @@ class DeleteShopToFavoriteProvider {
   Future<bool> getisDeleteShopTofavoriteresponse() async {
     String token = await loadToken();
     final response = await http.delete(
-      Uri.parse('$apiHead/clients/api/profile/favorite-stores/$storeUuid?api_token=$token'),
+      Uri.parse(
+          '$apiHead/clients/api/profile/favorite-stores/$storeUuid?api_token=$token'),
       headers: {'Accept': 'application/json'},
     );
-    print('url:: delete ' + '$apiHead/clients/api/profile/favorite-stores/$storeUuid?api_token=$token');
+    print('url:: delete ' +
+        '$apiHead/clients/api/profile/favorite-stores/$storeUuid?api_token=$token');
     if (response.statusCode == 204)
       return true;
     else {
@@ -1595,7 +1804,8 @@ class ProductDetailsProvider {
       }
     }
 
-    final response = await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
+    final response =
+        await http.get(Uri.parse(url), headers: {'Accept': 'application/json'});
     print('url:: get ' + url);
     if (response.statusCode == 200) {
       final Map<String, dynamic> productDetJson = json.decode(response.body);
@@ -1611,18 +1821,21 @@ class AddDeleteProductToFavoriteProvider {
   final String productUuid;
   final bool isLiked;
 
-  AddDeleteProductToFavoriteProvider({@required this.isLiked, @required this.productUuid});
+  AddDeleteProductToFavoriteProvider(
+      {@required this.isLiked, @required this.productUuid});
 
   Future<String> getisAddProductTofavoriteResponse() async {
     String token = await loadToken();
     final response = isLiked == false
         ? await http.post(
-            Uri.parse('$apiHead/clients/api/profile/favorite-assortments?api_token=$token'),
+            Uri.parse(
+                '$apiHead/clients/api/profile/favorite-assortments?api_token=$token'),
             body: {'assortment_uuid': productUuid},
             headers: {'Accept': 'application/json'},
           )
         : await http.delete(
-            Uri.parse('$apiHead/clients/api/profile/favorite-assortments/{$productUuid}?api_token=$token'),
+            Uri.parse(
+                '$apiHead/clients/api/profile/favorite-assortments/{$productUuid}?api_token=$token'),
             headers: {'Accept': 'application/json'},
           );
     print('❤️ getisAddProductTofavoriteResponse');
@@ -1648,11 +1861,15 @@ class ProfileProvider {
     String _token = await loadToken();
     _profileUrl = "$apiHead/clients/api/profile";
 
-    final respone = await http.delete(Uri.parse(_profileUrl),
-        headers: {'Accept': 'application/json', 'Authorization': 'Bearer $_token', 'Content-Type': 'application/json'});
+    final respone = await http.delete(Uri.parse(_profileUrl), headers: {
+      'Accept': 'application/json',
+      'Authorization': 'Bearer $_token',
+      'Content-Type': 'application/json'
+    });
     print('url:: delete ' + _profileUrl);
     if (respone.statusCode == 200) {
-      final Map<String, dynamic> deleteProfileREsponseJson = json.decode(respone.body);
+      final Map<String, dynamic> deleteProfileREsponseJson =
+          json.decode(respone.body);
       return DeleteProfileModel.fromJson(deleteProfileREsponseJson);
     } else {
       throw Exception("Error fetching delete profile response");
@@ -1669,8 +1886,8 @@ class ProfileProvider {
     }
     String token = await loadToken();
     _profileUrl = "$apiHead/clients/api/profile/push-tokens?api_token=$token";
-    final response =
-        await http.post(Uri.parse(_profileUrl), headers: {'Accept': 'application/json'}, body: {"token": pushToken});
+    final response = await http.post(Uri.parse(_profileUrl),
+        headers: {'Accept': 'application/json'}, body: {"token": pushToken});
     print('url:: post ' + _profileUrl);
     if (response.statusCode == 204) {
       return true;
@@ -1688,8 +1905,10 @@ class ProfileProvider {
       return false;
     }
     String token = await loadToken();
-    _profileUrl = "$apiHead/clients/api/profile/push-tokens/$pushToken?api_token=$token";
-    final response = await http.delete(Uri.parse(_profileUrl), headers: {'Accept': 'application/json'});
+    _profileUrl =
+        "$apiHead/clients/api/profile/push-tokens/$pushToken?api_token=$token";
+    final response = await http.delete(Uri.parse(_profileUrl),
+        headers: {'Accept': 'application/json'});
     print('url:: delete ' + _profileUrl);
     if (response.statusCode == 204) {
       return true;
@@ -1700,31 +1919,39 @@ class ProfileProvider {
 
   Future<String> getPurchasesSumResponse({@required int days}) async {
     String token = await loadToken();
-    String purchasesSumUrl = "$apiHead/clients/api/profile/purchases-sum?api_token=$token&days=$days";
-    final response = await http.get(Uri.parse(purchasesSumUrl), headers: {'Accept': 'application/json'});
+    String purchasesSumUrl =
+        "$apiHead/clients/api/profile/purchases-sum?api_token=$token&days=$days";
+    final response = await http.get(Uri.parse(purchasesSumUrl),
+        headers: {'Accept': 'application/json'});
     print('url:: get ' + purchasesSumUrl);
     if (response.statusCode == 200) {
       final Map<String, dynamic> purchasesSumJson = json.decode(response.body);
       return PurchasesSumModel.fromJson(purchasesSumJson).data.toString();
     } else {
-      throw Exception("Error fetching purchases sum. StatusCode: ${response.statusCode}");
+      throw Exception(
+          "Error fetching purchases sum. StatusCode: ${response.statusCode}");
     }
   }
 
   Future<ProfileModel> getProfileResponse() async {
     String token = await loadToken();
-    final response = await http
-        .get(Uri.parse('$apiHead/clients/api/profile?api_token=$token'), headers: {'Accept': 'application/json'});
-    print('url:: get получаем профиль 😎 ' + '$apiHead/clients/api/profile?api_token=$token');
+    final response = await http.get(
+        Uri.parse('$apiHead/clients/api/profile?api_token=$token'),
+        headers: {'Accept': 'application/json'});
+    print('url:: get получаем профиль 😎 ' +
+        '$apiHead/clients/api/profile?api_token=$token');
     if (response.statusCode == 200) {
       final Map<String, dynamic> profileJson = json.decode(response.body);
       ProfileModel profileModel = ProfileModel.fromJson(profileJson);
       if (profileModel.data.selectedStoreUserUuid != null) {
         log('🌎🌎🌎🌎🌎🌎 получаем профиль selected_store_user_uuid = ${profileModel.data.selectedStoreUserUuid}');
         log('🌎🌎🌎🌎🌎🌎 получаем профиль selectedStoreAddress = ${profileModel.data.selectedStoreAddress}');
-        await prefs.setString(SharedKeys.shopLogo, profileModel.data.image.thumbnails.the200X200);
-        await prefs.setString(SharedKeys.shopUuid, profileModel.data.selectedStoreUserUuid);
-        await prefs.setString(SharedKeys.shopAddress, profileModel.data.selectedStoreAddress);
+        await prefs.setString(
+            SharedKeys.shopLogo, profileModel.data.image.thumbnails.the200X200);
+        await prefs.setString(
+            SharedKeys.shopUuid, profileModel.data.selectedStoreUserUuid);
+        await prefs.setString(
+            SharedKeys.shopAddress, profileModel.data.selectedStoreAddress);
       } else {
         log('🌎🌎🌎🌎🌎🌎 в профиле нет выбранного магазина ❗️❗️❗️');
       }
@@ -1734,7 +1961,8 @@ class ProfileProvider {
       if (response.statusCode == 401) {
         return ProfileModel();
       } else {
-        throw Exception('Error fetching profile data. StatusCode: ${response.statusCode}');
+        throw Exception(
+            'Error fetching profile data. StatusCode: ${response.statusCode}');
       }
     }
   }
@@ -1772,35 +2000,48 @@ class ProfileUpdateDataProvider {
     String token = await loadToken();
     String phoneFromShared = await loadPhone();
 
-    final response = await http.put(Uri.parse("$apiHead/clients/api/profile?api_token=$token"), headers: {
-      'Accept': 'application/json'
-    }, body: {
-      if (selectedStoreUserUuid != null) "selected_store_user_uuid": selectedStoreUserUuid,
-      'phone': phone != null ? phone : phoneFromShared,
-      if (name != null) 'name': name,
-      if (appVersion != null) "app_version": appVersion,
-      if (consentToServiceNewsletter != null) "consent_to_service_newsletter": consentToServiceNewsletter.toString(),
-      if (consentToReceivePromotionalMailings != null)
-        'consent_to_receive_promotional_mailings': consentToReceivePromotionalMailings.toString(),
-      if (email != null) 'email': email,
-      if (sex != null) 'sex': sex,
-      if (birthdayDate != null) "birth_date": birthdayDate,
-      if (isAgreeWithDiverseFoodPromo != null)
-        "is_agree_with_diverse_food_promo": isAgreeWithDiverseFoodPromo.toString(),
-    });
-    log('☎️😊 url:: обновляем профиль put getProfileChangeRsponse ' + "$apiHead/clients/api/profile?api_token=$token");
+    final response = await http.put(
+        Uri.parse("$apiHead/clients/api/profile?api_token=$token"),
+        headers: {
+          'Accept': 'application/json'
+        },
+        body: {
+          if (selectedStoreUserUuid != null)
+            "selected_store_user_uuid": selectedStoreUserUuid,
+          'phone': phone != null ? phone : phoneFromShared,
+          if (name != null) 'name': name,
+          if (appVersion != null) "app_version": appVersion,
+          if (consentToServiceNewsletter != null)
+            "consent_to_service_newsletter":
+                consentToServiceNewsletter.toString(),
+          if (consentToReceivePromotionalMailings != null)
+            'consent_to_receive_promotional_mailings':
+                consentToReceivePromotionalMailings.toString(),
+          if (email != null) 'email': email,
+          if (sex != null) 'sex': sex,
+          if (birthdayDate != null) "birth_date": birthdayDate,
+          if (isAgreeWithDiverseFoodPromo != null)
+            "is_agree_with_diverse_food_promo":
+                isAgreeWithDiverseFoodPromo.toString(),
+        });
+    log('☎️😊 url:: обновляем профиль put getProfileChangeRsponse ' +
+        "$apiHead/clients/api/profile?api_token=$token");
     log('🌎🌎🌎🌎🌎🌎 обновляем профиль selected_store_user_uuid = $selectedStoreUserUuid');
-    if (selectedStoreUserUuid != null) print("*selected_store_user_uuid: $selectedStoreUserUuid");
+    if (selectedStoreUserUuid != null)
+      print("*selected_store_user_uuid: $selectedStoreUserUuid");
     print("*phone: ${phone != null ? phone : phoneFromShared}");
     if (name != null) print("*name: $name");
     if (appVersion != null) print("*app_version: $appVersion");
-    if (consentToServiceNewsletter != null) print("*consent_to_service_newsletter: $consentToServiceNewsletter");
+    if (consentToServiceNewsletter != null)
+      print("*consent_to_service_newsletter: $consentToServiceNewsletter");
     if (consentToReceivePromotionalMailings != null)
-      print("*consent_to_receive_promotional_mailings: $consentToReceivePromotionalMailings");
+      print(
+          "*consent_to_receive_promotional_mailings: $consentToReceivePromotionalMailings");
     if (email != null) print("*email: $email");
     if (sex != null) print("*sex: $sex");
     if (birthdayDate != null) print("*birth_date: $birthdayDate");
-    if (isAgreeWithDiverseFoodPromo != null) print("*is_agree_with_diverse_food_promo: $isAgreeWithDiverseFoodPromo");
+    if (isAgreeWithDiverseFoodPromo != null)
+      print("*is_agree_with_diverse_food_promo: $isAgreeWithDiverseFoodPromo");
 
     if (response.statusCode == 200) {
       log('✅ Профиль успешно обновлен!');
@@ -1816,45 +2057,57 @@ class ProfileUpdateDataProvider {
 class ShoppingListsProvider {
   Future<ShoppingListsModel> getShoppingListsResponse() async {
     String token = await loadToken();
-    final response = await http.get(Uri.parse('$apiHead/clients/api/profile/shopping-lists?api_token=$token'),
+    final response = await http.get(
+        Uri.parse(
+            '$apiHead/clients/api/profile/shopping-lists?api_token=$token'),
         headers: {'Accept': 'application/json'});
-    print('url:: Получение списков покупок get ' + '$apiHead/clients/api/profile/shopping-lists?api_token=$token');
+    print('url:: Получение списков покупок get ' +
+        '$apiHead/clients/api/profile/shopping-lists?api_token=$token');
     if (response.statusCode == 200) {
       final Map<String, dynamic> shoppingListsJson = json.decode(response.body);
       return ShoppingListsModel.fromJson(shoppingListsJson);
     } else {
-      throw Exception('Error fetching shopping lists data!. StatusCode: ${response.statusCode}');
+      throw Exception(
+          'Error fetching shopping lists data!. StatusCode: ${response.statusCode}');
     }
   }
 }
 
 //get shopping list details
 class ShoppingListDetailsProvider {
-  Future<ShoppingListDeatailsModel> getShoppingListDetailsResponse({@required String shoppingListUuid}) async {
+  Future<ShoppingListDeatailsModel> getShoppingListDetailsResponse(
+      {@required String shoppingListUuid}) async {
     String token = await loadToken();
     String shopUuid = await loadShopUuid();
-    String shoppingListDetailsUrl = "$apiHead/clients/api/profile/shopping-lists/$shoppingListUuid?api_token=$token";
+    String shoppingListDetailsUrl =
+        "$apiHead/clients/api/profile/shopping-lists/$shoppingListUuid?api_token=$token";
 
     if (shopUuid != null) {
       shoppingListDetailsUrl += "&store_uuid=$shopUuid";
     }
-    final response = await http.get(Uri.parse(shoppingListDetailsUrl), headers: {'Accept': 'application/json'});
+    final response = await http.get(Uri.parse(shoppingListDetailsUrl),
+        headers: {'Accept': 'application/json'});
     print('url:: get ' + shoppingListDetailsUrl);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> shoppingListDetailsJson = json.decode(response.body);
+      final Map<String, dynamic> shoppingListDetailsJson =
+          json.decode(response.body);
       return ShoppingListDeatailsModel.fromJson(shoppingListDetailsJson);
     } else {
-      throw Exception('Error fetching shopping lists data!. StatusCode: ${response.statusCode}');
+      throw Exception(
+          'Error fetching shopping lists data!. StatusCode: ${response.statusCode}');
     }
   }
 }
 
 //create new shopping list
 class CreateNewShoppingListProvider {
-  Future<bool> getCreateShoppingListResponse({@required String shoppingListName}) async {
+  Future<bool> getCreateShoppingListResponse(
+      {@required String shoppingListName}) async {
     String token = await loadToken();
-    final response = await http.post(Uri.parse('$apiHead/clients/api/profile/shopping-lists'),
-        headers: {"Accept": "application/json"}, body: {'name': shoppingListName, 'api_token': token});
+    final response = await http.post(
+        Uri.parse('$apiHead/clients/api/profile/shopping-lists'),
+        headers: {"Accept": "application/json"},
+        body: {'name': shoppingListName, 'api_token': token});
     print('url:: post ' + '$apiHead/clients/api/profile/shopping-lists');
     // print('name:: ' + shoppingListName);
     // print('api_token:: ' + token);
@@ -1871,10 +2124,12 @@ class CreateNewShoppingListProvider {
       List<double> assortmentsUuid,
       @required String shoppingListsUuid}) async {
     String token = await loadToken();
-    String _url = '$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}';
+    String _url =
+        '$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}';
 
     final response = await http.put(Uri.parse(_url),
-        headers: {"Accept": "application/json"}, body: {'name': shoppingListName, 'api_token': token});
+        headers: {"Accept": "application/json"},
+        body: {'name': shoppingListName, 'api_token': token});
     print('url:: put ' + _url);
     // print('name:: ' + shoppingListName);
     // print('api_token:: ' + token);
@@ -1888,13 +2143,16 @@ class CreateNewShoppingListProvider {
 
 // delete Shopping list
 class DeleteShoppingListProvider {
-  Future<bool> getDeleteShoppingListResponse({@required String shoppingListsUuid}) async {
+  Future<bool> getDeleteShoppingListResponse(
+      {@required String shoppingListsUuid}) async {
     String token = await loadToken();
     final response = await http.delete(
-      Uri.parse("$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}?api_token=$token"),
+      Uri.parse(
+          "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}?api_token=$token"),
       headers: {"Accept": "application/json"},
     );
-    print('url:: delete ' + "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}?api_token=$token");
+    print('url:: delete ' +
+        "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}?api_token=$token");
     if (response.statusCode == 200) {
       return true;
     } else {
@@ -1906,14 +2164,22 @@ class DeleteShoppingListProvider {
 //add product to shopping list
 class AddProductToShopingListProvider {
   Future<bool> getAddProducttoShoppingListRespone(
-      {@required String shoppingListsUuid, @required String assortmentUuid, @required int quantity}) async {
+      {@required String shoppingListsUuid,
+      @required String assortmentUuid,
+      @required int quantity}) async {
     String token = await loadToken();
     final response = await http.post(
-      Uri.parse("$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}/assortments"),
-      body: {"api_token": token, "assortment_uuid": assortmentUuid, "quantity": quantity.toString()},
+      Uri.parse(
+          "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}/assortments"),
+      body: {
+        "api_token": token,
+        "assortment_uuid": assortmentUuid,
+        "quantity": quantity.toString()
+      },
       headers: {"Accept": "application/json"},
     );
-    print('url:: post ' + "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}/assortments");
+    print('url:: post ' +
+        "$apiHead/clients/api/profile/shopping-lists/{$shoppingListsUuid}/assortments");
     // print('api_token:: ' + token);
     // print('assortment_uuid:: ' + assortmentUuid);
     // print('quantity:: ' + quantity.toString());
@@ -1928,7 +2194,8 @@ class AddProductToShopingListProvider {
 //Delete product to shopping list
 class DeleteProductToShopingListProvider {
   Future<bool> getDeleteProducttoShoppingListRespone(
-      {@required String shoppingListsUuid, @required String assortmentUuid}) async {
+      {@required String shoppingListsUuid,
+      @required String assortmentUuid}) async {
     String token = await loadToken();
     final response = await http.delete(
       Uri.parse(
@@ -1957,15 +2224,17 @@ class ShoppingCheckListProvider {
     print('url:: get ' +
         '$apiHead/clients/api/profile/receipts?api_token=$token&where[0][0]=refund_by_receipt_uuid&where[0][1]=is null&order_by[created_at]=desc');
     if (response.statusCode == 200) {
-      final Map<String, dynamic> shoppingCheckListJson = json.decode(response.body);
+      final Map<String, dynamic> shoppingCheckListJson =
+          json.decode(response.body);
       return ShoppingCheckListModel.fromJson(shoppingCheckListJson);
     } else {
       throw Exception('error fetching shopping checks list');
     }
   }
 
-  Future<List<ShoppingCheckListDataModel>> getShoppingCheckListForPaginationResponse(
-      {@required int currentPage}) async {
+  Future<List<ShoppingCheckListDataModel>>
+      getShoppingCheckListForPaginationResponse(
+          {@required int currentPage}) async {
     String token = await loadToken();
     final response = await http.get(
       Uri.parse(
@@ -1975,7 +2244,8 @@ class ShoppingCheckListProvider {
     print('url:: get ' +
         '$apiHead/clients/api/profile/receipts?api_token=$token&page=$currentPage&where[0][0]=refund_by_receipt_uuid&where[0][1]=is null&order_by[created_at]=desc');
     if (response.statusCode == 200) {
-      final Map<String, dynamic> shoppingCheckListJson = json.decode(response.body);
+      final Map<String, dynamic> shoppingCheckListJson =
+          json.decode(response.body);
       return ShoppingCheckListModel.fromJson(shoppingCheckListJson).data;
     } else {
       throw Exception('error fetching shopping checks list');
@@ -1991,9 +2261,11 @@ class LoyaltyCardsListProvider {
       Uri.parse('$apiHead/clients/api/profile/loyalty-cards?api_token=$token'),
       headers: {"Accept": "application/json"},
     );
-    print('url:: get ' + '$apiHead/clients/api/profile/loyalty-cards?api_token=$token');
+    print('url:: get ' +
+        '$apiHead/clients/api/profile/loyalty-cards?api_token=$token');
     if (response.statusCode == 200) {
-      final Map<String, dynamic> loyaltyCardsListJson = json.decode(response.body);
+      final Map<String, dynamic> loyaltyCardsListJson =
+          json.decode(response.body);
       return LoyaltyCardsListModel.fromJson(loyaltyCardsListJson);
     } else {
       throw Exception('error fetching lyalty card data');
@@ -2007,9 +2279,11 @@ class Asd<St> {}
 class CheckDetailsProvider {
   String _url = "";
 
-  Future<CheckDetailsModel> getCheckDetailsResponse({@required String receiptUuid}) async {
+  Future<CheckDetailsModel> getCheckDetailsResponse(
+      {@required String receiptUuid}) async {
     String token = await loadToken();
-    _url = '$apiHead/clients/api/profile/receipts/{$receiptUuid}?api_token=$token';
+    _url =
+        '$apiHead/clients/api/profile/receipts/{$receiptUuid}?api_token=$token';
     final response = await http.get(
       Uri.parse(_url),
       headers: {"Accept": "application/json"},
@@ -2028,10 +2302,12 @@ class CheckDetailsProvider {
 class CheckDetailsProductsProvider {
   String _url = "";
 
-  Future<CheckDetailsProductsModel> getCheckDetailsProductsResponse({@required String receiptUuid}) async {
+  Future<CheckDetailsProductsModel> getCheckDetailsProductsResponse(
+      {@required String receiptUuid}) async {
     String token = await loadToken();
 
-    _url = '$apiHead/clients/api/profile/receipts/{$receiptUuid}/lines?api_token=$token';
+    _url =
+        '$apiHead/clients/api/profile/receipts/{$receiptUuid}/lines?api_token=$token';
 
     final response = await http.get(
       Uri.parse(_url),
@@ -2039,7 +2315,8 @@ class CheckDetailsProductsProvider {
     );
     print('url:: get ' + _url);
     if (response.statusCode == 200) {
-      final Map<String, dynamic> checkDetailsProductsJson = json.decode(response.body);
+      final Map<String, dynamic> checkDetailsProductsJson =
+          json.decode(response.body);
       return CheckDetailsProductsModel.fromJson(checkDetailsProductsJson);
     } else {
       throw Exception('error fetching check details data');
@@ -2087,12 +2364,14 @@ class ReceiptsListProvider {
 
   Future<UniqueReceipts> getReceiptsUniqueSections() async {
     String token = await loadToken();
-    String _url = "$apiHead/clients/api/meal-receipts-unique-sections?api_token=$token";
+    String _url =
+        "$apiHead/clients/api/meal-receipts-unique-sections?api_token=$token";
     final response = await http.get(
       Uri.parse(_url),
       headers: {"Accept": "application/json"},
     );
-    print('url get getReceiptsUniqueSections 🍓 рецепты only категории:: ' + _url);
+    print('url get getReceiptsUniqueSections 🍓 рецепты only категории:: ' +
+        _url);
     if (response.statusCode == 200) {
       final receiptsListJson = jsonDecode(response.body);
       return UniqueReceipts.fromMap(receiptsListJson);
@@ -2104,7 +2383,8 @@ class ReceiptsListProvider {
   Future<SingleRecipeDataModel> getSingleReceipt(String uuid) async {
     String token = await loadToken();
     String _url = "$apiHead/clients/api/meal-receipts/$uuid/?api_token=$token";
-    final response = await http.get(Uri.parse(_url), headers: {"Accept": "application/json"});
+    final response = await http
+        .get(Uri.parse(_url), headers: {"Accept": "application/json"});
     print('url:: get getSingleReceipt 🍓 рецепт 1:: ' + _url);
     if (response.statusCode == 200) {
       final receiptsListJson = jsonDecode(response.body);
@@ -2117,9 +2397,13 @@ class ReceiptsListProvider {
   Future<bool> toggleLikeStory(String uuid, bool isFavorite) async {
     //это вообще лайк для сториса рецепта а не избранное
     String token = await loadToken();
-    String _url = "$apiHead/clients/api/meal-receipts/$uuid/reaction?api_token=$token";
+    String _url =
+        "$apiHead/clients/api/meal-receipts/$uuid/reaction?api_token=$token";
     final response = await http.post(Uri.parse(_url),
-        headers: {"Accept": "application/json", "Content-Type": "application/json"},
+        headers: {
+          "Accept": "application/json",
+          "Content-Type": "application/json"
+        },
         body: jsonEncode({"is_positive": isFavorite}));
     print('url:: post addToFavorites 🍓 лайк сториса рецепта:: ' + _url);
     // print('is_positive:: ' + isFavorite.toString());
@@ -2136,17 +2420,20 @@ class AddOrDeleteReceiptsToFavoriteProvider {
   final String mealReceiptUuid;
   final bool isFavorite;
 
-  AddOrDeleteReceiptsToFavoriteProvider({@required this.isFavorite, @required this.mealReceiptUuid});
+  AddOrDeleteReceiptsToFavoriteProvider(
+      {@required this.isFavorite, @required this.mealReceiptUuid});
 
   Future<String> addOrDeleteReceiptsToFavoriteResponse() async {
     String token = await loadToken();
     final response = isFavorite == true
         ? await http.delete(
-            Uri.parse('$apiHead/clients/api/profile/favorite-meal-receipts/$mealReceiptUuid?api_token=$token'),
+            Uri.parse(
+                '$apiHead/clients/api/profile/favorite-meal-receipts/$mealReceiptUuid?api_token=$token'),
             headers: {'Accept': 'application/json'},
           )
         : await http.post(
-            Uri.parse('$apiHead/clients/api/profile/favorite-meal-receipts?api_token=$token'),
+            Uri.parse(
+                '$apiHead/clients/api/profile/favorite-meal-receipts?api_token=$token'),
             body: {'meal_receipt_uuid': mealReceiptUuid},
             headers: {'Accept': 'application/json'},
           );
