@@ -5,19 +5,19 @@ import 'package:smart/core/constants/source.dart';
 import 'package:smart/core/constants/text_styles.dart';
 
 class AppAlert {
-  static OverlayEntry _overlayEntry;
+  static OverlayEntry? _overlayEntry;
 
   static Future<void> show({
-    BuildContext context,
-    String message,
+    required BuildContext context,
+    required String message,
     String messageTap = '',
     int sec = 10,
     String svgName = 'newCustomAlert.svg',
     bool isPushToContactsWidget = false,
-    Function() funcPushToContactsWidgetOnTap,
+    VoidCallback? funcPushToContactsWidgetOnTap,
   }) async {
     if (_overlayEntry != null) {
-      _overlayEntry.remove();
+      _overlayEntry!.remove();
       _overlayEntry = null;
     }
 
@@ -34,15 +34,25 @@ class AppAlert {
             decoration: BoxDecoration(
               borderRadius: BorderRadius.circular(10),
               color: Colors.white,
-              boxShadow: [BoxShadow(color: grey04, offset: Offset(1, 2), blurRadius: 2, spreadRadius: 0)],
+              boxShadow: [
+                BoxShadow(
+                    color: grey04,
+                    offset: Offset(1, 2),
+                    blurRadius: 2,
+                    spreadRadius: 0)
+              ],
             ),
-            padding: EdgeInsets.symmetric(vertical: heightRatio(size: 11, context: context), horizontal: widthRatio(size: 16, context: context)),
+            padding: EdgeInsets.symmetric(
+                vertical: heightRatio(size: 11, context: context),
+                horizontal: widthRatio(size: 16, context: context)),
             child: Row(
               children: [
                 Container(
                   height: heightRatio(size: 49, context: context),
                   width: widthRatio(size: 49, context: context),
-                  padding: EdgeInsets.symmetric(horizontal: widthRatio(size: 8, context: context), vertical: heightRatio(size: 12, context: context)),
+                  padding: EdgeInsets.symmetric(
+                      horizontal: widthRatio(size: 8, context: context),
+                      vertical: heightRatio(size: 12, context: context)),
                   decoration: BoxDecoration(
                     color: newRedDark,
                     borderRadius: BorderRadius.circular(10),
@@ -69,17 +79,30 @@ class AppAlert {
                       ? RichText(
                           text: TextSpan(
                             text: message,
-                            style: appLabelTextStyle(fontSize: heightRatio(size: 13, context: context), color: newBlackLight, height: 1.2),
+                            style: appLabelTextStyle(
+                                fontSize:
+                                    heightRatio(size: 13, context: context),
+                                color: newBlackLight,
+                                height: 1.2),
                             children: <TextSpan>[
                               TextSpan(
                                 text: messageTap,
-                                style: appLabelTextStyle(fontSize: heightRatio(size: 13, context: context), color: newRedDark, height: 1.2),
-                                recognizer: TapGestureRecognizer()..onTap = funcPushToContactsWidgetOnTap,
+                                style: appLabelTextStyle(
+                                    fontSize:
+                                        heightRatio(size: 13, context: context),
+                                    color: newRedDark,
+                                    height: 1.2),
+                                recognizer: TapGestureRecognizer()
+                                  ..onTap = funcPushToContactsWidgetOnTap,
                               ),
                             ],
                           ),
                         )
-                      : Text(message, style: appLabelTextStyle(fontSize: heightRatio(size: 13, context: context), color: newBlackLight, height: 1.2)),
+                      : Text(message,
+                          style: appLabelTextStyle(
+                              fontSize: heightRatio(size: 13, context: context),
+                              color: newBlackLight,
+                              height: 1.2)),
                 ),
               ],
             ),
@@ -88,10 +111,10 @@ class AppAlert {
       ),
     );
 
-    Overlay.of(context).insert(_overlayEntry);
+    Overlay.of(context).insert(_overlayEntry!);
 
     Future.delayed(Duration(seconds: sec), () {
-      _overlayEntry.remove();
+      _overlayEntry?.remove();
       _overlayEntry = null;
     });
   }
